% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{add_stepcolumn}
\alias{add_stepcolumn}
\title{Add step column and fitrank}
\usage{
add_stepcolumn(myparframe, tol = 1)
}
\arguments{
\item{myparframe}{a parframe}

\item{tol}{integer for steps.}
}
\value{
the parframe with columns fitrank and step
}
\description{
Add step column and fitrank
}
