% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{as.data.table.reportYaml}
\alias{as.data.table.reportYaml}
\title{Title}
\usage{
\method{as.data.table}{reportYaml}(reportYaml)
}
\arguments{
\item{reportYaml}{}
}
\description{
Title
}
\seealso{
Other petab select: 
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_LRT}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
