% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_applyPaginate}
\alias{cf_applyPaginate}
\title{Get list of paginated plots}
\usage{
cf_applyPaginate(pl)
}
\arguments{
\item{pl}{(not yet facetted) ggplot}

\item{paginateInfo}{output from \link{cf_paginateInfo}}
}
\value{
list of ggplots
}
\description{
use ggforce
}
\examples{
# adapted from ggforce examples
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ cut:clarity, ncol = 3, nrow = 3, page = 4)
plotlist <- cf_applyPaginate(pl)
plotlist[[1]]
plotlist[[5]]
n_pages(pl)
length(plotlist)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
