% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_as.parframe}
\alias{cf_as.parframe}
\title{Better as_parframe}
\usage{
cf_as.parframe(x, sort.by = "value", ...)
}
\arguments{
\item{x}{}

\item{sort.by}{}

\item{...}{}
}
\value{

}
\description{
Adds AIC and BIC automatically, adds stepcolumn automatically
}
