% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_normIndiv.R
\name{cf_normL2_indiv}
\alias{cf_normL2_indiv}
\title{Fast normL2 alternative}
\usage{
cf_normL2_indiv(
  data,
  prd0,
  errmodel = NULL,
  est.grid,
  fixed.grid,
  times = NULL,
  attr.name = "data",
  fixed.conditions = NULL
)
}
\arguments{
\item{data}{}

\item{prd0}{}

\item{errmodel}{}

\item{est.grid}{}

\item{times}{}

\item{attr.name}{}

\item{fixed.conditions}{}

\item{fix.grid}{}
}
\value{
objective function
}
\description{
Fast normL2 alternative
}
\author{
Svenja Kemmer
}
