% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_outputFigure}
\alias{cf_outputFigure}
\title{Output figures}
\usage{
cf_outputFigure(
  pl,
  filename = NULL,
  width = 16,
  height = 10,
  scale = 1,
  units = c("cm", "mm", "in")[1],
  dpi = 300,
  limitsize = TRUE,
  device = NULL,
  heightrel = NULL,
  FLAGFuture = TRUE,
  FLAGoverwrite = TRUE,
  ...
)
}
\arguments{
\item{pl}{plot like ggplot, potentially created with facet_grid_paginate or list of ggplots}

\item{filename, width, height, scale, units, dpi, limitsize, device, ...}{see [ggplot2::ggsave()]}

\item{FLAGFuture}{Export asynchronously with the future-package}
}
\value{
nothing
}
\description{
Handle paginate automatically, save asynchronously with future
}
\details{
# importFrom ggplot2 plot_dev parse_dpi plot_dim
}
