% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_parf_getPars}
\alias{cf_parf_getPars}
\title{Select parameter columns of parframe}
\usage{
cf_parf_getPars(parf)
}
\arguments{
\item{parf}{parframe}

\item{parameters}{keine ahnung mehr?}
}
\description{
Select parameter columns of parframe
}
