% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_parf_getStepRepresentatives}
\alias{cf_parf_getStepRepresentatives}
\title{Title}
\usage{
cf_parf_getStepRepresentatives(parf, tol = 1)
}
\arguments{
\item{parf}{}

\item{tol}{}
}
\value{

}
\description{
Title
}
