% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_parf_parNames}
\alias{cf_parf_parNames}
\title{get Parameter names of a parframe}
\usage{
cf_parf_parNames(x)
}
\arguments{
\item{x}{parframe}
}
\description{
get Parameter names of a parframe
}
