% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_parf_rbindlist}
\alias{cf_parf_rbindlist}
\title{Rbind some parframes}
\usage{
cf_parf_rbindlist(parflist)
}
\arguments{
\item{parflist}{list(parframes)}
}
\value{
parframe
}
\description{
Rbind some parframes
}
\examples{
parflist <- list(cf_parframe(data.frame(value = 1, a = 1, b = 2), metanames = "value"),
                 cf_parframe(data.frame(value = 1, a = 2, parameterSetId = "par2"), metanames = c("value", "parameterSetId")))
parflist <- list(cf_parframe(data.frame(value = 1, a = 1), metanames = "value"),
                 cf_parframe(data.frame(value = 1, a = 2, parameterSetId = "par2"), metanames = c("value", "parameterSetId")))
cf_parf_rbindlist(parflist)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
