% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_parframe}
\alias{cf_parframe}
\title{Improved version of parframe}
\usage{
cf_parframe(
  x = NULL,
  parameters = NULL,
  metanames = NULL,
  obj.attributes = NULL,
  tol = 1
)
}
\arguments{
\item{x}{}

\item{parameters}{}

\item{metanames}{}

\item{obj.attributes}{}

\item{tol}{}
}
\value{

}
\description{
Fits coerces to data.frame and guesses metanames. Adds stepcolumn automatically
}
