% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_normIndiv.R
\name{cf_PRD_indiv}
\alias{cf_PRD_indiv}
\title{PRD_indiv alternative}
\usage{
cf_PRD_indiv(prd0, est.grid, fixed.grid)
}
\arguments{
\item{prd0}{}

\item{est.grid}{}

\item{fix.grid}{}
}
\value{

}
\description{
PRD_indiv alternative
}
\author{
Svenja Kemmer
}
