% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_profile_getOpenProfiles}
\alias{cf_profile_getOpenProfiles}
\title{Title}
\usage{
cf_profile_getOpenProfiles(profiles, value_column = "data")
}
\arguments{
\item{profiles}{dMod::profile}

\item{value_name}{Choose from numerical attributes of obj, e.g. "value", "data", "prior" ...}
}
\value{
data.table(whichPar, profileDirection, isOpen)
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{profiles}
