% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_profile_getProfilesOptimumBelowFit}
\alias{cf_profile_getProfilesOptimumBelowFit}
\title{getProfiles with optimum below the fit$argument}
\usage{
cf_profile_getProfilesOptimumBelowFit(profiles)
}
\arguments{
\item{profiles}{parframe}
}
\value{
data.table
}
\description{
Sometimes, a profile finds a better optimum
Get only profiles of parameters which show this behaviour
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
