% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_profile_plotPathsAffected}
\alias{cf_profile_plotPathsAffected}
\title{Title}
\usage{
cf_profile_plotPathsAffected(
  profiles,
  tol = 0.1,
  FLAGnormalizeYParameters = TRUE,
  nrow = 4,
  ncol = 5,
  ggCallback = list(),
  ...
)
}
\arguments{
\item{profiles}{profiles to plotPaths}

\item{tol}{}

\item{page}{pagination page}
}
\value{
paginated ggplot
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
