% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cf_profile_prepareAffectedPaths}
\alias{cf_profile_prepareAffectedPaths}
\title{Prepare parameters affected by a profile}
\usage{
cf_profile_prepareAffectedPaths(
  profiles,
  tol = 0.1,
  FLAGnormalizeYParameters = TRUE,
  SDcols = c("value", "constraint", "PARAMETER1", "PARAMETER2", "PARVALUE1",
    "PARVALUE2")
)
}
\arguments{
\item{profiles}{Profiles from dMod::profile}

\item{tol}{tolerance to measure if a parameter is affected}

\item{FLAGnormalizeYParameters}{Let them all pass through the same point: subtract the value at constraint = 0}
}
\value{
data.table(value, constraint, PARAMETER1, PARAMETER2, PARVALUE1, PARVALUE2)]
}
\description{
Its not helpful to plot all paths in plotPaths,
but only of those parameters which are affected
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
