% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\docType{data}
\name{cfcolors}
\alias{cfcolors}
\title{Colors
copied and adapted from dMod}
\format{
An object of class \code{character} of length 136.
}
\usage{
cfcolors
}
\description{
Colors
copied and adapted from dMod
}
\examples{
ggplot(data.frame(x = 1:42,color = factor(cfcolors[1:42], unique(cfcolors[1:42])))) + 
  geom_tile(aes(x=x, y = 1, fill = color)) + 
  scale_color_manual(values= setNames(nm = cfcolors),
  aesthetics = c("color", "fill"))
}
\keyword{datasets}
