% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\docType{data}
\name{cfcolorsFULL}
\alias{cfcolorsFULL}
\title{Colors Copied from dMod}
\format{
An object of class \code{list} of length 3.
}
\usage{
cfcolorsFULL
}
\description{
Colors Copied from dMod
}
\keyword{datasets}
