% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cfggplot}
\alias{cfggplot}
\title{cfggplot}
\usage{
cfggplot(data = NULL, mapping = aes(), FLAGbold = TRUE)
}
\arguments{
\item{data, mapping}{see ?ggplot}

\item{FLAGbold}{Make text bold faced}
}
\value{
ggplot
}
\description{
cfggplot
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
