% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{cfgrab}
\alias{cfgrab}
\title{Grab a base plot into grid object}
\usage{
cfgrab()
}
\value{
gtree, can be exported with [ggplot2::ggsave()]
}
\description{
Grab a base plot into grid object
}
