% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{check_clusterTimeStamp}
\alias{check_clusterTimeStamp}
\title{Check for the time stamp}
\usage{
check_clusterTimeStamp(FLAGforcePurge = FALSE)
}
\arguments{
\item{FLAGforcePurge}{Force the removal of a time stamp}
}
\description{
If no recent login-stamp is available or if stamp is invalidated (> 1 hour){
if stamp is invalidated: {remove stamp}
stop, force creation of new stamp}
}
\seealso{
Other Cluster login helpers: 
\code{\link{write_clusterTimeStamp}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster login helpers}
