% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{dco_expandPatterns}
\alias{dco_expandPatterns}
\title{Expand}
\usage{
dco_expandPatterns(dco, pe)
}
\arguments{
\item{dco}{dco without meta columns}

\item{pe}{pe. If meta$emtaInformation$petab_columns specification present, this information will be used to generate new columns

Example:
pe$meta$emtaInformation$petab_columns = list(replicateId = "replicate_gel_experiment", conditionId = "cellline_stimulus")
will split
a) the replicateId column into three columns called "replicate", "gel" and "experiment" which are added to the dco
b) the conditionId column into cellline and stimulus}
}
\value{
dco
}
\description{
Expand
}
\seealso{
Other dco: 
\code{\link{petab_joinDCO}()},
\code{\link{petab_mutateDCO}()},
\code{\link{petab_unjoinDCO}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{dco}
