% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{dMod_files}
\alias{dMod_files}
\title{Consistent dMod filenames}
\usage{
dMod_files(path, identifier = "")
}
\arguments{
\item{path}{}

\item{identifier}{}
}
\value{

}
\description{
Consistent dMod filenames
}
