% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools_SB.R
\name{dmod_obs_2_petab_obs}
\alias{dmod_obs_2_petab_obs}
\title{dMod Observables to PEtab converter}
\usage{
dmod_obs_2_petab_obs(obs, param_identifier = c("scale", "offset"))
}
\arguments{
\item{obs}{\link{eqnvec}}

\item{param_identifier}{vector of strings identifying parameters that should
be replaced in the PEtab format. The passed phrases must be appear in all 
parameters that should be replaced and in none that not. Per default it is
assumed that there are (only) \code{scale} and \code{offset} parameters
identifiable by these exact phrases}
}
\value{
data.table formatted as required by PEtab
}
\description{
Function that takes an eqnvec object from dMod and converts it
to the desired PEtab format including renaming of scaling and offset-
parameters.
}
\examples{

## generate 'eqnvec' object with four parameters variably formatted 
observ <- eqnvec(test = "parafoo + meter_bar * cyberglarb(zorbhyper)")

## define what the parameter identiyfyers are
identifiers <- c("foo", "bar", "glarb", "zorb")

## execute function
out <- dmod_obs_2_petab_obs(obs = observ, param_identifier = identifiers)

}
\author{
Severin Bang <severin.bang@physik.uni-freiburg.de>
}
