% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{dMod_readMstrust}
\alias{dMod_readMstrust}
\title{Title}
\usage{
dMod_readMstrust(path, identifier = NULL)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
