% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{dMod_saveL1}
\alias{dMod_saveL1}
\title{Title}
\usage{
dMod_saveL1(L1, path, identifier = "1", FLAGoverwrite = FALSE)
}
\arguments{
\item{path}{}

\item{identifier}{}

\item{FLAGoverwrite}{}

\item{fit}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
