% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{dMod_saveMstrust}
\alias{dMod_saveMstrust}
\title{Title}
\usage{
dMod_saveMstrust(fit, path, identifier = "1", FLAGoverwrite = FALSE)
}
\arguments{
\item{fit}{parlist or parframe, will be coerced to parframe with \code{\link[=cf_as.parframe]{cf_as.parframe()}}}

\item{path, identifier}{result will be written to \code{\link[=dMod_files]{dMod_files()}}(path, identifier)$mstrust}

\item{FLAGoverwrite}{TRUE or FALSE}
}
\value{
fit as parframe, invisibly
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
