% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{dMod_saveProfiles}
\alias{dMod_saveProfiles}
\title{Consistently save profiles}
\usage{
dMod_saveProfiles(profs, path)
}
\arguments{
\item{path}{}

\item{profiles}{}
}
\value{

}
\description{
Consistently save profiles
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
