% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getEventInfo}
\alias{getEventInfo}
\title{Title}
\usage{
getEventInfo(eventList)
}
\arguments{
\item{eventList}{\code{\link[dMod:eventList]{dMod::eventList()}}}
}
\value{
data.table()
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
