% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{getEXgrid}
\alias{getEXgrid}
\title{Get parameter formulas from dMod trafo}
\usage{
getEXgrid(est.grid, fixed.grid)
}
\arguments{
\item{est.grid}{as output by dMod::getParGrids()[\link{1}]}

\item{fixed.grid}{as output by dMod::getParGrids()[\link{2}]}
}
\value{
merge of est.grid and fixed.grid
}
\description{
Get parameter formulas from dMod trafo
}
\seealso{
Other Parameter wrangling: 
\code{\link{petab_getParameterFormulas}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()},
\code{\link{strsplit2}()}
}
\author{
Svenja Kemmer
}
\concept{Parameter wrangling}
