% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{getFacets}
\alias{getFacets}
\title{Title}
\usage{
getFacets(fp, type)
}
\arguments{
\item{fp, type}{see code of \code{\link[=getPaginateInfo]{getPaginateInfo()}}}
}
\value{

}
\description{
Title
}
\examples{
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ cut:clarity, ncol = 3, nrow = 3, page = 4)
fp <- pl$facet$params
type <- "facet_grid_paginate"
getFacets(fp, type)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
