% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{getPaginateInfo}
\alias{getPaginateInfo}
\title{get the pagination info from a plot}
\usage{
getPaginateInfo(pl)
}
\arguments{
\item{pl}{ggplot}
}
\value{

}
\description{
get the pagination info from a plot
}
\examples{
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ cut:clarity, ncol = 3, nrow = 3, page = 4)
getPaginateInfo(pl)
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate(color ~ ., nrow = 2, ncol = 2, page = 1)
getPaginateInfo(pl)
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_grid_paginate( ~ color, nrow = 2, ncol = 2, page = 1)
getPaginateInfo(pl)
pl <- ggplot(diamonds) +
  geom_point(aes(carat, price), alpha = 0.1) +
  facet_wrap_paginate( ~ color, nrow = 2, ncol = 2, page = 1)
getPaginateInfo(pl)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
