% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getUnitInfo}
\alias{getUnitInfo}
\title{Assemble unitInfo}
\usage{
getUnitInfo(unitName = NULL, unitKind = NULL, exponent = NULL)
}
\arguments{
\item{unitName}{character vector}

\item{unitKind}{list of character vectors}

\item{exponent}{list of numeric vectors}
}
\value{
data.table(unitName, unitKind, exponent)
}
\description{
Some default units, some others
}
\examples{
getUnitInfo()
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneParameter}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
