% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getUnitInfoSimple}
\alias{getUnitInfoSimple}
\title{Assemble unitInfo using only time}
\usage{
getUnitInfoSimple(
  unitName = NULL,
  unitKind = NULL,
  exponent = NULL,
  scale = NULL,
  multiplier = NULL
)
}
\arguments{
\item{unitName}{character vector}

\item{unitKind}{list of character vectors}

\item{exponent}{list of numeric vectors}
}
\value{
data.table(unitName, unitKind, exponent)
}
\description{
Some default units, some others
}
\examples{
getUnitInfo()
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameterNum}()},
\code{\link{sbml_addOneParameterSym}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpeciesNum}()},
\code{\link{sbml_addOneSpeciesSym}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Svenja Kemmer
}
\concept{SBML export}
