% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-indiv2Classic.R
\name{indiv2Classic_gridlist2cond.grid}
\alias{indiv2Classic_gridlist2cond.grid}
\title{Title}
\usage{
indiv2Classic_gridlist2cond.grid(gridlist)
}
\arguments{
\item{gridlist}{gridlist}
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
