% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-indiv2Classic.R
\name{indiv2Classic_trafo}
\alias{indiv2Classic_trafo}
\title{Title}
\usage{
indiv2Classic_trafo(trafo, cg)
}
\arguments{
\item{trafo}{base trafo}

\item{cg}{output of \code{\link[=indiv2Classic_gridlist2cond.grid]{indiv2Classic_gridlist2cond.grid()}}}
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
