% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{inputFileChanged}
\alias{inputFileChanged}
\title{Determine if intput file is newer than output file}
\usage{
inputFileChanged(filename_input, filename_output)
}
\arguments{
\item{filename_input, filename_output}{file path}
}
\value{
TRUE if inputfile is newer than outputfile or outputfile does not exist
}
\description{
Determine if intput file is newer than output file
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
