% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{inverse_scale}
\alias{inverse_scale}
\title{Title}
\usage{
inverse_scale(parameterScale)
}
\arguments{
\item{parameterScale}{vector with entries "lin", "log", "log10"}
}
\value{
vector with entries "lin", "exp", "10^"
}
\description{
Title
}
\seealso{
Other scales: 
\code{\link{lin}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{scales}
