% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_knecht.R
\name{knechtStatusMessage}
\alias{knechtStatusMessage}
\title{Title}
\usage{
knechtStatusMessage(FLAGjobDone, FLAGjobPurged, FLAGjobRecover)
}
\arguments{
\item{FLAGjobDone}{}

\item{FLAGjobPurged}{}

\item{FLAGjobRecover}{}
}
\value{

}
\description{
Title
}
\seealso{
Other knecht: 
\code{\link{pd_knecht_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_knecht_L1_mstrust}()},
\code{\link{pd_knecht_L1}()},
\code{\link{pd_knecht_mstrust}()},
\code{\link{pd_knecht_profile}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{knecht}
