% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{L1_getModelCandidates}
\alias{L1_getModelCandidates}
\title{Title}
\usage{
L1_getModelCandidates(L1Scan)
}
\arguments{
\item{L1Scan}{L1-parframe Result from pd_cluster_L1}
}
\value{
Matrix indicating the model structure of L1 models
}
\description{
Title
}
