% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{L1_getParameterFormulaInjection}
\alias{L1_getParameterFormulaInjection}
\title{Create L1 Trafo, depending on parameterScale}
\usage{
L1_getParameterFormulaInjection(pe, parameterId_base)
}
\arguments{
\item{pe}{}

\item{parameterId_base}{vector of parameterId which should be L1'd.}
}
\value{
data.table(parameterId, parameterFormula, trafoType = "L1")
}
\description{
\itemize{
\item For logpars: par * L1_par
\item For linpars: par + L1_par
}
}
\examples{
pe <- petab_exampleRead("05")
L1_getParameterFormulaInjection(pe, c("kcat", "E"))
}
\seealso{
Other petab L1: 
\code{\link{pe_L1_addL1ParsToExperimentalCondition}()},
\code{\link{pe_L1_createL1Problem}()},
\code{\link{pe_L1_updateParameterFormulaInjection}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab L1}
