% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{lin}
\alias{lin}
\title{Identity}
\usage{
lin(x)
}
\arguments{
\item{x}{}
}
\value{
x
}
\description{
Identity
}
\seealso{
Other scales: 
\code{\link{inverse_scale}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{scales}
