% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{pars2parframe}
\alias{pars2parframe}
\title{Turn a pars vector into a single-row parframe}
\usage{
pars2parframe(pars, parameterSetId = "Base", obj = NULL)
}
\arguments{
\item{pars}{setNames(outervalues, parnames)}

\item{parameterSetId}{Identifier for this parameterset}

\item{obj}{Objective function like normL2}
}
\value{
parframe
}
\description{
Turn a pars vector into a single-row parframe
}
\examples{
obj <- function(pars) {list(value = sum(pars^2))}
pars = c(a = 1, b = 2)
pars2parframe(pars, "base", obj)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
