% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_addObjPrior}
\alias{pd_addObjPrior}
\title{Title}
\usage{
pd_addObjPrior(pd, parameters, FLAGuseNominalValueAsCenter)
}
\arguments{
\item{pd}{}

\item{parameters}{}
}
\description{
Title
}
\examples{
parameters <- c("k_233", "k_234", "k_244")
}
