% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_cluster_L1_fitUnbiasedEachOnce}
\alias{pd_cluster_L1_fitUnbiasedEachOnce}
\title{Fit model on cluster}
\usage{
pd_cluster_L1_fitUnbiasedEachOnce(
  pd,
  .outputFolder,
  n_startsPerNode = 16 * 3,
  n_cores = 64,
  identifier = "L1UBSingle",
  FLAGforcePurge = FALSE,
  passwdEnv = Sys.getenv("hurensohn"),
  machine = "cluster"
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_startsPerNode}{}

\item{n_cores}{}

\item{n_nodes}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on cluster
}
\seealso{
Other Cluster: 
\code{\link{clusterStatusMessage}()},
\code{\link{pd_cluster_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_cluster_L1}()},
\code{\link{pd_cluster_mstrust}()},
\code{\link{pd_cluster_profile}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster}
