% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_cluster_L1_fitUnbiasedKNECHT}
\alias{pd_cluster_L1_fitUnbiasedKNECHT}
\title{Fit model on cluster}
\usage{
pd_cluster_L1_fitUnbiasedKNECHT(
  pd,
  .outputFolder,
  n_startsPerNode = 24 * 2,
  machine_identifier = "ruprecht1",
  FLAGforcePurge = FALSE
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_startsPerNode}{}

\item{n_nodes}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on cluster
}
\seealso{
Other Cluster: 
\code{\link{clusterStatusMessage}()},
\code{\link{pd_cluster_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_cluster_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_cluster_L1}()},
\code{\link{pd_cluster_mstrust}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster}
