% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_cluster_profile}
\alias{pd_cluster_profile}
\title{Run profiles on cluster}
\usage{
pd_cluster_profile(
  pd,
  .outputFolder,
  FLAGforcePurge = FALSE,
  FLAGfixParsOnBoundary = TRUE,
  n_cores = 64,
  profpars = names(pd$pars),
  passwdEnv = Sys.getenv("hurensohn"),
  machine = "cluster"
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{FLAGforcePurge}{}

\item{FLAGfixParsOnBoundary}{Fix parameters which went to the boundary. Don't fit and don't
use for reoptimization. A bit heuristic, but \code{\link[dMod:profile]{dMod::profile()}} does not support boundaries.
Therefore, if those parameters were not fixed, one potentially not start from the "optimum"}

\item{n_cores}{cores for the cluster to use per node}

\item{profpars}{I suggest to use either 1. names(pd$pars) or 2. hard code the result
from \code{\link[=pd_profile_getParsNotYetProfiled]{pd_profile_getParsNotYetProfiled()}}}
}
\description{
Will profile the parameters stored in pd$pars
}
\seealso{
Other Cluster: 
\code{\link{clusterStatusMessage}()},
\code{\link{pd_cluster_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_cluster_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_cluster_L1}()},
\code{\link{pd_cluster_mstrust}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster}
