% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fitObsPars}
\alias{pd_fitObsPars}
\title{Fit only obsPars}
\usage{
pd_fitObsPars(pd, FLAGoverwrite = FALSE, iterlim = 500)
}
\arguments{
\item{pd}{}

\item{FLAGoverwrite}{TRUE or FALSE}
}
\value{
pd with pd$result$pd$base_obsParsFitted
}
\description{
Fit only obsPars
}
\seealso{
Other pd fitting: 
\code{\link{pd_mstrust}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd fitting}
