% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_knecht.R
\name{pd_knecht_L1}
\alias{pd_knecht_L1}
\title{Fit model on knecht}
\usage{
pd_knecht_L1(
  pd,
  .outputFolder,
  n_nodes = 6,
  lambdas = 10^(seq(log10(1e-04), log10(100), length.out = n_nodes * 16 - 1)),
  identifier = "L1",
  FLAGforcePurge = FALSE,
  passwdEnv = Sys.getenv("hurensohn"),
  machine = "knecht"
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_nodes}{}

\item{n_startsPerNode}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on knecht
}
\seealso{
Other knecht: 
\code{\link{knechtStatusMessage}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_knecht_L1_mstrust}()},
\code{\link{pd_knecht_mstrust}()},
\code{\link{pd_knecht_profile}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{knecht}
