% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_knecht.R
\name{pd_knecht_L1_mstrust}
\alias{pd_knecht_L1_mstrust}
\title{Fit model on knecht}
\usage{
pd_knecht_L1_mstrust(
  pd,
  .outputFolder,
  nstarts = 240,
  n_nodes = length(pd$pe$meta$L1$parameters_L1reference) * 1.5,
  lambdas = 10^(seq(log10(0.001), log10(1000), length.out = n_nodes)),
  identifier = "L1mstrust",
  FLAGforcePurge = FALSE,
  passwdEnv = Sys.getenv("hurensohn"),
  machine = "knecht"
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_nodes}{}

\item{n_startsPerNode}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on knecht
}
\seealso{
Other knecht: 
\code{\link{knechtStatusMessage}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_knecht_L1}()},
\code{\link{pd_knecht_mstrust}()},
\code{\link{pd_knecht_profile}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{knecht}
