% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_knecht.R
\name{pd_knecht_mstrust}
\alias{pd_knecht_mstrust}
\title{Fit model on knecht}
\usage{
pd_knecht_mstrust(
  pd,
  .outputFolder,
  nStartsPerCore = 4,
  identifier = "mstrust",
  FLAGforcePurge = FALSE,
  opt.parameter_startpoints = "sample",
  iterlim = 500,
  machine
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_startsPerNode}{}

\item{n_nodes}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on knecht
}
\seealso{
Other knecht: 
\code{\link{knechtStatusMessage}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_knecht_L1_mstrust}()},
\code{\link{pd_knecht_L1}()},
\code{\link{pd_knecht_profile}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{knecht}
