% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_knecht.R
\name{pd_knecht_profile}
\alias{pd_knecht_profile}
\title{Run profiles on knecht}
\usage{
pd_knecht_profile(
  pd,
  .outputFolder,
  FLAGforcePurge = FALSE,
  FLAGfixParsOnBoundary = TRUE,
  profpars = names(pd$pars),
  passwdEnv = Sys.getenv("hurensohn"),
  machine = "knecht"
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{FLAGforcePurge}{}

\item{FLAGfixParsOnBoundary}{Fix parameters which went to the boundary. Don't fit and don't
use for reoptimization. A bit heuristic, but \code{\link[dMod:profile]{dMod::profile()}} does not support boundaries.
Therefore, if those parameters were not fixed, one potentially not start from the "optimum"}

\item{profpars}{I suggest to use either 1. names(pd$pars) or 2. hard code the result
from \code{\link[=pd_profile_getParsNotYetProfiled]{pd_profile_getParsNotYetProfiled()}}}
}
\value{

}
\description{
Will profile the parameters stored in pd$pars
}
\seealso{
Other knecht: 
\code{\link{knechtStatusMessage}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachMstrust}()},
\code{\link{pd_knecht_L1_fitUnbiasedEachOnce}()},
\code{\link{pd_knecht_L1_mstrust}()},
\code{\link{pd_knecht_L1}()},
\code{\link{pd_knecht_mstrust}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{knecht}
