% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_createPetabFromUnbiasedFit}
\alias{pd_L1_createPetabFromUnbiasedFit}
\title{Title}
\usage{
pd_L1_createPetabFromUnbiasedFit(pd, L1modelId, filename = NULL)
}
\arguments{
\item{pd}{}

\item{L1modelId}{L1 model Id you want to export}

\item{filename}{}
}
\value{

}
\description{
Title
}
\examples{
pd_L1_createPetabFromUnbiasedFit(pd, L1modelId = 4) # if 4 is for example the parsimonious model you want to profile
}
\seealso{
Other L1: 
\code{\link{pd_L1_getObjBounds}()},
\code{\link{pd_L1_getUnbiasedParframeAllCandidates}()},
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotDevianceVsLambda}()},
\code{\link{pd_L1_plotIsConvergedVsLambda}()},
\code{\link{pd_L1_plotParValueVsLambda}()},
\code{\link{pd_L1_plotValueVsLambda}()},
\code{\link{pd_L1_preparePlottingData}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
