% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_getObjBounds}
\alias{pd_L1_getObjBounds}
\title{Title}
\usage{
pd_L1_getObjBounds(pd, tol = 1e-04)
}
\arguments{
\item{pd}{pd with pd$result$L1}

\item{tol}{Tolerance}
}
\value{
data.table(lambdaL1 = lambda, Ntotal = Total parameters which were L1'd, NFree = Remaining free parameters, df = degrees of freedom, chis = The quantile value which corresponds to significance level alpha = 0.5)
}
\description{
Title
}
\seealso{
Other L1: 
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotDevianceVsLambda}()},
\code{\link{pd_L1_plotIsConvergedVsLambda}()},
\code{\link{pd_L1_plotValueVsLambda}()},
\code{\link{pd_L1_preparePlottingData}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
