% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_plotDevianceVsLambda}
\alias{pd_L1_plotDevianceVsLambda}
\title{Title}
\usage{
pd_L1_plotDevianceVsLambda(pd, ...)
}
\arguments{
\item{pd}{with pd$result$L1}

\item{...}{Arguments going to cf_outputFigure}
}
\value{
ggplot
}
\description{
Title
}
\seealso{
Other L1: 
\code{\link{pd_L1_getObjBounds}()},
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotIsConvergedVsLambda}()},
\code{\link{pd_L1_plotParValueVsLambda}()},
\code{\link{pd_L1_plotValueVsLambda}()},
\code{\link{pd_L1_preparePlottingData}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
