% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_plotIsConvergedVsLambda}
\alias{pd_L1_plotIsConvergedVsLambda}
\title{Title}
\usage{
pd_L1_plotIsConvergedVsLambda(pd, ...)
}
\arguments{
\item{pd}{}

\item{...}{}
}
\value{
ggplot
}
\description{
Title
}
\seealso{
Other L1: 
\code{\link{pd_L1_createPetabFromUnbiasedFit}()},
\code{\link{pd_L1_getObjBounds}()},
\code{\link{pd_L1_getUnbiasedParframeAllCandidates}()},
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotDevianceVsLambda}()},
\code{\link{pd_L1_plotParValueVsLambda}()},
\code{\link{pd_L1_plotValueVsLambda}()},
\code{\link{pd_L1_preparePlottingData}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
