% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_plotValueVsLambda}
\alias{pd_L1_plotValueVsLambda}
\title{Title}
\usage{
pd_L1_plotValueVsLambda(pd, ...)
}
\arguments{
\item{pd}{}

\item{...}{}
}
\value{
ggplot
}
\description{
Title
}
\seealso{
Other L1: 
\code{\link{pd_L1_getObjBounds}()},
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotDevianceVsLambda}()},
\code{\link{pd_L1_plotIsConvergedVsLambda}()},
\code{\link{pd_L1_preparePlottingData}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
