% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_L1_preparePlottingData}
\alias{pd_L1_preparePlottingData}
\title{Prepare data for some L1 plots}
\usage{
pd_L1_preparePlottingData(pd)
}
\arguments{
\item{pd}{with pd$result$L1}
}
\value{
data.table(parameterId,parameterType,isL1Parameter,lambda,value,iterations,converged,muValue,estValue,estDeviance)
}
\description{
Prepare data for some L1 plots
}
\seealso{
Other L1: 
\code{\link{pd_L1_getObjBounds}()},
\code{\link{pd_L1_getUnbiasedValues}()},
\code{\link{pd_L1_plotDevianceVsLambda}()},
\code{\link{pd_L1_plotIsConvergedVsLambda}()},
\code{\link{pd_L1_plotValueVsLambda}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{L1}
