% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collect}
\alias{pd_parf_collect}
\title{Collect parameters for simultaneous prediction or plotting}
\usage{
pd_parf_collect(
  pd,
  opt.base = pd_parf_opt.base(include = TRUE, parameterSetId = "Base"),
  opt.mstrust = pd_parf_opt.mstrust(include = TRUE, fitrankRange = 1:20, tol = 1),
  opt.profile = list(include = FALSE, rows = "profile_endpoints", parameters = NULL)
)
}
\arguments{
\item{opt.profile}{}
}
\value{
pd with pd$parf
}
\description{
Collect parameters for simultaneous prediction or plotting
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
