% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collectL1}
\alias{pd_parf_collectL1}
\title{Title}
\usage{
pd_parf_collectL1(pd, rows = NULL)
}
\arguments{
\item{rows}{numeric vector of row indices. NULL keeps all rows}
}
\value{
\code{\link[dMod:parframe]{dMod::parframe()}}
}
\description{
Title
}
\seealso{
Other parframe handling: 
\code{\link{pd_parf_collectMstrust}()},
\code{\link{pd_parf_collectPars}()},
\code{\link{pd_parf_collectProfile}()},
\code{\link{pd_parf_collect}()},
\code{\link{pd_updateEstPars}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{parframe handling}
