% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collectMstrust}
\alias{pd_parf_collectMstrust}
\title{Title}
\usage{
pd_parf_collectMstrust(pd, fitrankRange = 1:15, tol = 1)
}
\arguments{
\item{fitrankRange}{}
}
\value{
\code{\link[dMod:parframe]{dMod::parframe()}}
}
\description{
Title
}
\seealso{
Other parframe handling: 
\code{\link{pd_parf_collectPars}()},
\code{\link{pd_parf_collectProfile}()},
\code{\link{pd_parf_collect}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{parframe handling}
