% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collectPars}
\alias{pd_parf_collectPars}
\title{Title}
\usage{
pd_parf_collectPars(pd, parameterSetId = "Base")
}
\arguments{
\item{pd}{}
}
\value{
\code{\link[dMod:parframe]{dMod::parframe()}}
}
\description{
Title
}
\seealso{
Other parframe handling: 
\code{\link{pd_parf_collectMstrust}()},
\code{\link{pd_parf_collectProfile}()},
\code{\link{pd_parf_collect}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{parframe handling}
