% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collectProfile}
\alias{pd_parf_collectProfile}
\title{Title}
\usage{
pd_parf_collectProfile(
  pd,
  rows = c("profile_endpoints", "optimum"),
  parameters = NULL
)
}
\arguments{
\item{parameters}{}
}
\value{

}
\description{
Title
}
