% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_pars_getFixedOnBoundary}
\alias{pd_pars_getFixedOnBoundary}
\title{Get parameters which are fixed on the boundary}
\usage{
pd_pars_getFixedOnBoundary(pd, tol = 0.01)
}
\arguments{
\item{pd}{pd with pd$pars from a fit}

\item{tol}{tolerance how close each parameter on the est-scale should be to the boundary in order to be fixed}
}
\value{
vector of fixed parameters
}
\description{
Get parameters which are fixed on the boundary
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
