% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{pd_petabSelect_collectParameters}
\alias{pd_petabSelect_collectParameters}
\title{Collect Parameters from PD}
\usage{
pd_petabSelect_collectParameters(pd)
}
\arguments{
\item{pd}{a pd}
}
\value{
named list(fixedPar = value, estimatedPar = "estimate")
}
\description{
Collect Parameters from PD
}
\examples{
pd <- petab_exampleRead("01", "pd")
pd_petabSelect_collectParameters(pd)
}
\seealso{
Other petab select: 
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
