% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{pd_petabSelect_collectReportYamlCriteria}
\alias{pd_petabSelect_collectReportYamlCriteria}
\title{Collect AIC and BIC from pd}
\usage{
pd_petabSelect_collectReportYamlCriteria(pd)
}
\arguments{
\item{pd}{a pd}
}
\value{
list(AIC = numeric(1L), BIC = numeric(1L))
}
\description{
Chooses LL based on availability
\enumerate{
\item If mstrust is available, take from mstrust
\item else take from "base" (which might be a single fit) CAREFUL: base might be worse than base_obsParsFitted, as base is not necessarily fitted
\item else take from "base_obsParsFitted"
\item else take from pd$pars (which might be unfitted)
}
}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
