% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_plot}
\alias{pd_plot}
\title{First version of plotting a pd}
\usage{
pd_plot(
  pd,
  ...,
  page = 1,
  nrow = 3,
  ncol = 4,
  filename = NULL,
  width = 29.7,
  height = 21,
  scale = 1,
  units = "cm"
)
}
\arguments{
\item{pd}{A pd}

\item{...}{Going to \code{\link[dMod:plotCombined]{dMod::plotCombined()}}}

\item{page, nrow, ncol}{Going to \code{\link[ggforce:facet_wrap_paginate]{ggforce::facet_wrap_paginate()}}}

\item{filename, width, height, scale, units}{Going to \code{\link[conveniencefunctions:cf_outputFigure]{conveniencefunctions::cf_outputFigure()}}}
}
\value{
ggplot
}
\description{
Predicts with "pars" and "times"
}
\examples{
pdx <- petab_exampleRead("01", "pd")
pd_plot(pdx)
}
\seealso{
Other Plotting: 
\code{\link{pd_predictAndPlot2}()},
\code{\link{pd_predictAndPlot}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Plotting}
