% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_plotParsParallelLines2}
\alias{pd_plotParsParallelLines2}
\title{Plot resulting parameters from mstrust}
\usage{
pd_plotParsParallelLines2(
  pd,
  stepMax = 3,
  filename = NULL,
  i,
  ggCallback = NULL,
  ...
)
}
\arguments{
\item{pd}{pd with field pd$result$mstrust}

\item{stepMax}{maximum step of waterfall}

\item{filename}{for plotting}

\item{i}{subset the data.table on their names c("parameterId", "parameterType", "fitrank", "step", "stepsize",
"index", "value", "converged", "iterations", "estValue")}

\item{ggCallback}{for plotting}

\item{...}{output to \code{\link[conveniencefunctions:cf_outputFigure]{conveniencefunctions::cf_outputFigure()}}}
}
\value{
ggplot
}
\description{
This plot is inspired by the Plot from the Tutorial on modeling by Villaverde
}
\seealso{
Other plotting: 
\code{\link{pd_plotParsParallelLines}()},
\code{\link{pd_plotProfile}()},
\code{\link{petab_plotHelpers_aeslist}()},
\code{\link{petab_plotHelpers_meanMeasurementsValue}()},
\code{\link{petab_plotHelpers_parameterOrder}()},
\code{\link{petab_plotHelpers_variableOrder}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotting}
