% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_predictAndPlot}
\alias{pd_predictAndPlot}
\title{Title}
\usage{
pd_predictAndPlot(
  pd,
  i,
  opt.base = pd_parf_opt.base(),
  opt.mstrust = pd_parf_opt.mstrust(),
  NFLAGsubsetType = c(none = 0, strict = 1, keepInternal = 2)[2],
  FLAGsummarizeProfilePredictions = TRUE,
  nrow = 3,
  ncol = 4,
  scales = "free",
  page = 1,
  n.breaks = 5,
  filename = NULL,
  width = 29.7,
  height = 21,
  scale = 1,
  units = "cm"
)
}
\arguments{
\item{i}{"data.table i" to subset data and predictions}

\item{FLAGsummarizeProfilePredictions}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
