% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_profile}
\alias{pd_profile}
\title{Title}
\usage{
pd_profile(
  pd,
  .outputFolder,
  FLAGfixParsOnBoundary = FALSE,
  whichPar = pd_profile_getParsNotYetProfiled(pd = pd, .outputFolder = .outputFolder,
    FLAGreturnVector = TRUE),
  ...
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{FLAGfixParsOnBoundary}{}

\item{profpars}{}
}
\description{
Title
}
