% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_profile_getParsNotYetProfiled}
\alias{pd_profile_getParsNotYetProfiled}
\title{Determine which pars were not yet profiled}
\usage{
pd_profile_getParsNotYetProfiled(pd, .outputFolder, FLAGreturnVector = FALSE)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{FLAGreturnVector}{}
}
\value{
either vector of p^arameters or output printed to console
}
\description{
Determine which pars were not yet profiled
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
