% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_rdsfile}
\alias{pd_rdsfile}
\title{Get the rds file path in one simple command}
\usage{
pd_rdsfile(pd)
}
\arguments{
\item{pd}{}
}
\value{
path of the rds file
}
\description{
Get the rds file path in one simple command
}
\seealso{
Other pd: 
\code{\link{pd_fitMstrust}()},
\code{\link{pd_fit}()},
\code{\link{pd_mstrust}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd}
