% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_removeProfiles}
\alias{pd_removeProfiles}
\title{Title}
\usage{
pd_removeProfiles(pd)
}
\arguments{
\item{pd}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{}
