% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{pdIndiv_getBaseScales}
\alias{pdIndiv_getBaseScales}
\title{Title}
\usage{
pdIndiv_getBaseScales(cg, scalesOuter)
}
\arguments{
\item{cg}{condition.grid}

\item{scalesOuter}{vector of scales of outer parameters c(CS_name = scale)}
}
\value{
vector of scales of base parameters c(BASE_name = scale)
}
\description{
Title
}
\examples{
cg <- data.table(structure(list(conditionId = c("C1", "C2"), conditionName = c("C1", "C2"), kcat = c("kcat_C1", "kcat_C2"), kon = c("kon_C1", "kon_C2"), koff = c(0.1, 0.2), observableParameter1_obsE = c("offset_E", "offset_E"), observableParameter1_obsS = c("offset_S", "offset_S_C2"), noiseParameter1_obsE = c("sigmaRel_obsE", "sigmaRel_obsE"), noiseParameter1_obsES = c("sigma_obsES", "sigma_obsES"), noiseParameter1_obsP = c("sigma_obsP", "sigma_obsP"), noiseParameter1_obsS = c("sigma_obsS", "sigma_obsS_C2"), noiseParameter2_obsE = c("sigmaAbs_obsE", "sigmaAbs_obsE"), E = c("E", "E"), ES = c("ES", "ES"), P = c("P", "P"), S = c("S", "S"), cytoplasm = c("cytoplasm", "cytoplasm")), row.names = c(NA, -2L), class = c("data.frame")))
scalesOuter <- c(E = "log10", ES = "log10", kcat_C1 = "log10", kcat_C2 = "log10", offset_E = "log10", offset_S = "log10", offset_S_C2 = "log10", P = "log10", S = "log10", sigma_obsES = "log10", sigma_obsP = "log10", sigma_obsS = "log10", sigma_obsS_C2 = "log10", sigmaAbs_obsE = "log10", sigmaRel_obsE = "log10", kon_C1 = "log10", kon_C2 = "log10", cytoplasm = "lin")
pdIndiv_getBaseScales(cg, scalesOuter)
}
\seealso{
Other SBML Import: 
\code{\link{pdIndiv_getParameters_conditionGrid}()},
\code{\link{pdIndiv_initializeGridlist}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML Import}
