% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{pdIndiv_initializeGridlist}
\alias{pdIndiv_initializeGridlist}
\title{Title}
\usage{
pdIndiv_initializeGridlist(mycondition.grid)
}
\arguments{
\item{mycondition.grid}{}
}
\description{
Title
}
\seealso{
Other SBML Import: 
\code{\link{pdIndiv_getBaseScales}()},
\code{\link{pdIndiv_getParameters_conditionGrid}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML Import}
