% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pdIndiv_updateControls}
\alias{pdIndiv_updateControls}
\title{Quickly set "controls" amd update all high-level functions}
\usage{
pdIndiv_updateControls(
  pd,
  optionsOde = NULL,
  optionsSens = NULL,
  objtimes = NULL
)
}
\arguments{
\item{objtimes}{}
}
\description{
Quickly set "controls" amd update all high-level functions
}
\examples{
pd <- pdIndiv_updateControls(pd, 
                             optionsOde  = list(method = "lsoda", rtol = 1e-12, atol = 1e-12, maxsteps = 10000),
                             optionsSens = list(method = "lsodes", rtol = 1e-12, atol = 1e-12, maxsteps = 10000))
saveRDS(pd,pd_rdsfile(pd))
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
