% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pdlist.R
\name{pdlist_petab_plotHelpers_aeslist}
\alias{pdlist_petab_plotHelpers_aeslist}
\title{Slightly modified default aesthetics for multimodel plotting}
\usage{
pdlist_petab_plotHelpers_aeslist(
  x = ~time,
  y = ~measurement,
  color = ~modelId,
  fill = ~modelId,
  linetype = ~parameterSetId,
  ymin = ~measurementmin,
  ymax = ~measurementmax,
  shape = ~modelId,
  ...
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{color}{}

\item{fill}{}

\item{linetype}{}

\item{ymin}{}

\item{ymax}{}

\item{shape}{}

\item{...}{}
}
\value{

}
\description{
Slightly modified default aesthetics for multimodel plotting
}
\seealso{
Other pdlist: 
\code{\link{pdlist_predictAndPlot2}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pdlist}
