% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pdlist.R
\name{pdlist_predictAndPlot2}
\alias{pdlist_predictAndPlot2}
\title{Plot with list of pds}
\usage{
pdlist_predictAndPlot2(
  pdlist,
  FLAGreturnPlotData = FALSE,
  nrow = 4,
  ncol = 5,
  aeslist = pdlist_petab_plotHelpers_aeslist(),
  ggCallback = list(),
  opt.gg = list(ribbonAlpha = 0.2),
  ...
)
}
\arguments{
\item{pdlist}{named list of pds}

\item{FLAGreturnPlotData}{return list(data.tables) which go into plotting instead of ggplot}

\item{nrow}{}

\item{ncol}{}

\item{aeslist}{list of aestheticss}

\item{ggCallback}{list(ggplot2 calls), e.g. list(labs(title = "bla"), scale_y_log10())}

\item{opt.gg}{}

\item{...}{Arguments going to pd_predictAndPlot2}
}
\value{
ggplot or plotData = list(dplot,pplot,pplotRibbon)
}
\description{
Plot with list of pds
}
\seealso{
Other pdlist: 
\code{\link{pdlist_petab_plotHelpers_aeslist}()}

Other Plotting: 
\code{\link{pd_predictAndPlot2}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Plotting}
\concept{pdlist}
