% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{pe_L1_addL1ParsToExperimentalCondition}
\alias{pe_L1_addL1ParsToExperimentalCondition}
\title{Add parameter columns to experimentalCondition}
\usage{
pe_L1_addL1ParsToExperimentalCondition(
  pe,
  parameterId_base,
  conditionSpecL1_reference
)
}
\arguments{
\item{pe}{petab with non-Null pe$experimentalCondition$L1Spec column which serves a) subsetting and b) creating parameter names}

\item{parameterId_base}{parameterIds to be L1'd}

\item{conditionSpecL1_reference}{Vector of entries referring to L1Spec which are deemed the base condition}
}
\value{
petab with updated experimentalCondition: L1Spec is removed, L1 parameter columns are added
}
\description{
Add parameter columns to experimentalCondition
}
\examples{
pe <- petab_exampleRead("04")
pe$experimentalCondition$L1Spec <- pe$experimentalCondition$conditionId
pe_L1_addL1ParsToExperimentalCondition(pe, parameterId_base = c("kcat", "E"), conditionSpecL1_reference = "C1")
}
\seealso{
Other petab L1: 
\code{\link{L1_getParameterFormulaInjection}()},
\code{\link{pe_L1_createL1Problem}()},
\code{\link{pe_L1_updateParameterFormulaInjection}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab L1}
