% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{pe_L1_createL1Problem}
\alias{pe_L1_createL1Problem}
\title{From a petab, create a petab which encodes an L1 problem}
\usage{
pe_L1_createL1Problem(
  pe,
  parameterId_base,
  conditionSpecL1_reference,
  j_conditionSpecL1 = conditionId
)
}
\arguments{
\item{pe}{petab}

\item{parameterId_base}{parameterIds to be L1'd}

\item{conditionSpecL1_reference}{Vector of entries referring to L1Spec which are deemed the base condition}

\item{j_conditionSpecL1}{data.table-j argument to create L1Spec within pe$experimentalCondition. E.g. conditionId (as symbol, will be substitute-eval'd)}
}
\value{
\code{\link[=petab]{petab()}} with L1 parameters
}
\description{
From a petab, create a petab which encodes an L1 problem
}
\examples{
pe <- petab_exampleRead("04")
pe_L1_createL1Problem(pe, c("kcat", "E"), conditionSpecL1_reference = "C1", j_conditionSpecL1 = conditionId)
}
\seealso{
Other petab L1: 
\code{\link{L1_getParameterFormulaInjection}()},
\code{\link{pe_L1_addL1ParsToExperimentalCondition}()},
\code{\link{pe_L1_updateParameterFormulaInjection}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab L1}
