% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{pe_L1_updateParameterFormulaInjection}
\alias{pe_L1_updateParameterFormulaInjection}
\title{Title}
\usage{
pe_L1_updateParameterFormulaInjection(pe, parameterId_base)
}
\arguments{
\item{pe}{}

\item{parameterId_base}{vector of parameterId which should be L1'd.}
}
\value{
petab with updated parameterFormulaInjection, see \code{\link[=L1_getParameterFormulaInjection]{L1_getParameterFormulaInjection()}}
}
\description{
Title
}
\examples{
pe <- petab_exampleRead("05")
pe_L1_updateParameterFormulaInjection(pe, c("kcat", "E"))
}
\seealso{
Other petab L1: 
\code{\link{L1_getParameterFormulaInjection}()},
\code{\link{pe_L1_addL1ParsToExperimentalCondition}()},
\code{\link{pe_L1_createL1Problem}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab L1}
