% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{pe_qNLME_updateParameterFormulaInjection}
\alias{pe_qNLME_updateParameterFormulaInjection}
\title{Title}
\usage{
pe_qNLME_updateParameterFormulaInjection(pe, parameterId_base)
}
\arguments{
\item{pe}{}

\item{parameterId_base}{vector of parameterId which should be qNLME'd.}
}
\value{
petab with updated parameterFormulaInjection, see \code{\link[=qNLME_getParameterFormulaInjection]{qNLME_getParameterFormulaInjection()}}
}
\description{
Title
}
\examples{
pe <- petab_exampleRead("05")
pe_qNLME_updateParameterFormulaInjection(pe, c("kcat", "E"))
}
\seealso{
Other petab qNLME: 
\code{\link{qNLME_getParameterFormulaInjection}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab qNLME}
