% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{pepy_sample_parameter_startpoints}
\alias{pepy_sample_parameter_startpoints}
\title{Sample parameter start points}
\usage{
pepy_sample_parameter_startpoints(
  pe,
  n_starts = 100L,
  seed = 1L,
  FLAGincludeCurrent = TRUE
)
}
\arguments{
\item{n_starts}{number of starting points}

\item{FLAGincludeCurrent}{Include values of pe$parameters in returned start parameters}
}
\value{
parframe
}
\description{
Uses Petab.py, is based on pe$parameters
}
\examples{
n_starts = 100L
seed = 1L
FLAGincludeCurrent = TRUE
}
\seealso{
Other pepy: 
\code{\link{petab_python_reinstall}()},
\code{\link{petab_python_setup}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pepy}
