% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab}
\alias{petab}
\title{Collector function for petab files}
\usage{
petab(
  model = NULL,
  experimentalCondition = NULL,
  measurementData = NULL,
  observables = NULL,
  parameters = NULL,
  meta = NULL,
  ...
)
}
\arguments{
\item{model}{\code{\link[dMod:eqnlist]{dMod::eqnlist()}}}

\item{observables}{see \code{\link[=petab_observables]{petab_observables()}}}

\item{parameters}{see \code{\link[=petab_parameters]{petab_parameters()}}}

\item{condition}{see \code{\link[=petab_condition]{petab_condition()}}}

\item{measurements}{see \code{\link[=petab_measurements]{petab_measurements()}}}
}
\value{
list of the input arguments
}
\description{
Collector function for petab files
}
\examples{
# [ ] Todo
# readPetab(filename)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
