% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_applyObservableTransformation}
\alias{petab_applyObservableTransformation}
\title{Title}
\usage{
petab_applyObservableTransformation(pe)
}
\arguments{
\item{pe}{}
}
\value{
pe
}
\description{
Title
}
\seealso{
Other measurementData: 
\code{\link{petab_applyInverseObservableTransformation}()},
\code{\link{petab_duplicateControls}()},
\code{\link{petab_fixErrorModel}()},
\code{\link{petab_measurementData}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{measurementData}
