% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_combine}
\alias{petab_combine}
\title{Combine two petabs}
\usage{
petab_combine(
  pe1,
  pe2,
  NFLAGconflict = c(stop = 0, use_pe1 = 1, use_pe2 = 2)[2]
)
}
\arguments{
\item{pe1, pe2}{petabs}

\item{NFLAGconflict}{one of c("stop" = 0, "use_pe1" = 1, "use_pe2" = 2)}
}
\value{
pe
}
\description{
\itemize{
\item Row-bind the relevant tables
\item If conflicts occur, inform user about it (e.g. different column names in experimentalCondition)
}
}
\seealso{
Other Combine petabs: 
\code{\link{petab_combine_experimentalCondition}()},
\code{\link{petab_combine_measurementData}()},
\code{\link{petab_combine_observables}()},
\code{\link{petab_combine_parameters}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Combine petabs}
