% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_combine_observables}
\alias{petab_combine_observables}
\title{Title}
\usage{
petab_combine_observables(
  o1,
  o2,
  NFLAGconflict = c(stop = 0, use_pe1 = 1, use_pe2 = 2)[2]
)
}
\arguments{
\item{o2}{}
}
\description{
Title
}
\seealso{
Other Combine petabs: 
\code{\link{petab_combine_experimentalCondition}()},
\code{\link{petab_combine_measurementData}()},
\code{\link{petab_combine_parameters}()},
\code{\link{petab_combine}()}
}
\concept{Combine petabs}
