% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_create_parameter_df}
\alias{petab_create_parameter_df}
\title{Original petab function didn't work}
\usage{
petab_create_parameter_df(pe, observableParameterScale = "log10")
}
\arguments{
\item{pe}{petab without a parameter_df}

\item{observableParameterScale}{"lin", "log" or "log10"}
}
\value{
\code{\link[=petab_parameters]{petab_parameters()}} data.table
}
\description{
Original petab function didn't work
}
\examples{
pe <- petab_exampleRead("01", "pe")
pe$parameters <- NULL
petab_create_parameter_df(pe)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
