% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_createObjPrior}
\alias{petab_createObjPrior}
\title{Create the prior function specified in parameters_df}
\usage{
petab_createObjPrior(pe, FLAGuseNominalValueAsCenter)
}
\arguments{
\item{FLAGuseNominalCenter}{}
}
\value{

}
\description{
Create the prior function specified in parameters_df
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
