% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_createqNLMEProblem}
\alias{petab_createqNLMEProblem}
\title{Title}
\usage{
petab_createqNLMEProblem(pe, priorStrength, Nsub, seed = 1)
}
\arguments{
\item{pe}{}

\item{priorStrength}{}

\item{Nsub}{}

\item{seed}{if not NULL, subjects will be sampled based on objectivePriorParameters}
}
\value{

}
\description{
Does not work yet with observable parameters
}
\examples{
# see example 09
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{}
