% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{petab_dModmodel2PE}
\alias{petab_dModmodel2PE}
\title{Write PE from dMod objects}
\usage{
petab_dModmodel2PE(
  ODEmodel,
  obsFun,
  errormodel,
  data,
  bestfit,
  trafo,
  estGrid,
  fixedGrid,
  eventList,
  lb = 6.14e-06,
  ub = 162754.8
)
}
\arguments{
\item{ODEmodel}{}

\item{obsFun}{}

\item{errormodel}{}

\item{data}{DT with obligatory columns name, time, value, sigma, condition}

\item{trafo}{}

\item{eventList}{}

\item{parameters}{}
}
\value{
pe: list with PEtab content that can be exported with writePetab()
}
\description{
Write PE from dMod objects
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameter}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Svenja Kemmer
}
\concept{SBML export}
