% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_dput}
\alias{petab_dput}
\title{dput for petab variables}
\usage{
petab_dput(pe, variable)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}}}

\item{variable}{(single) character denoting a column, e.g. "observableId"}
}
\value{
dput output
}
\description{
dput for petab variables
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
