% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_duplicateControls}
\alias{petab_duplicateControls}
\title{Duplicate control data points}
\usage{
petab_duplicateControls(pe, i, conditionMapping)
}
\arguments{
\item{i}{logical expression operating on the DC (not DCO) indicating which data points are to be duplicated}

\item{conditionMapping}{data.table(conditionId, conditionIdNew) where conditionId is the control condition and conditionIdNew is the condition where to the data point shall be copied}
}
\value{
pe with updated
}
\description{
Mainly for plotting, e.g. such that mean-lines start at zero
}
\seealso{
Other plotData: 
\code{\link{petab_plotData}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotData}
