% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{petab_exampleRead}
\alias{petab_exampleRead}
\title{Read an example}
\usage{
petab_exampleRead(exampleName, object = c("pe", "pd")[1])
}
\arguments{
\item{exampleName}{(partial match of) example name}

\item{object}{"pe", "pd" or "dir"}
}
\value{
pd or pe
}
\description{
Read an example
}
\examples{
pd <- petab_exampleRead("01-EnzymeKinetics", "pd")
}
\seealso{
Other Example functions: 
\code{\link{petab_exampleList}()},
\code{\link{petab_examplePath}()},
\code{\link{petab_select_examplePath}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Example functions}
