% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_experimentalCondition}
\alias{petab_experimentalCondition}
\title{Constructor for Conditions}
\usage{
petab_experimentalCondition(conditionId, conditionName = NA, ...)
}
\arguments{
\item{conditionId}{character Condition ID}

\item{conditionName}{character Condition ID for plotting}

\item{...}{parameterOrSpeciesOrCompartmentId1. Numeric (value) or string (different parameterId)}
}
\value{

}
\description{
https://petab.readthedocs.io/en/stable/documentation_data_format.html#condition-table
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
