% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_files}
\alias{petab_files}
\title{List petab files}
\usage{
petab_files(filename, FLAGTestCase = FALSE, FLAGreturnList = FALSE)
}
\arguments{
\item{filename}{"path/to/modelname.petab". Will generate filenames like
"path/to/modelname/model_modelname.xml"}

\item{FLAGTestCase}{generate TestCases filename}

\item{FLAGreturnList}{return list or vector?}
}
\value{
list or character vector of file paths
}
\description{
List petab files
}
\examples{
petab_files("Models/Example")
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
