% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_files_default}
\alias{petab_files_default}
\title{Get default filenames}
\usage{
petab_files_default(modelname, path)
}
\arguments{
\item{modelname}{}

\item{path}{}
}
\value{
list of file.paths
}
\description{
Get default filenames
}
\examples{
petab_files_default("modelname", "path")
}
\seealso{
Other petab_files: 
\code{\link{petabYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab_files}
