% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_files_fromYaml}
\alias{petab_files_fromYaml}
\title{Read list of filenames from a petabYaml}
\usage{
petab_files_fromYaml(filename)
}
\arguments{
\item{filename}{file.path to petab.yaml}
}
\value{
list of filenames
}
\description{
Read list of filenames from a petabYaml
}
\examples{
filename <- file.path(petab_examplePath("01"), "petab.yaml")
petab_files_fromYaml(filename)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab files}
