% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_fixErrorModel}
\alias{petab_fixErrorModel}
\title{Replace error parameters in measurementData by sensible estimates}
\usage{
petab_fixErrorModel(pe)
}
\arguments{
\item{pe}{}
}
\value{
pe
}
\description{
\enumerate{
\item Calculates sd per time, observableId and conditionId.
\item Then the mean of these sd's is taken per observableId
}
}
\details{
All calculations are done on the est-scale of each observable

Assumes constant error model (per observable)
}
\seealso{
Other measurementData: 
\code{\link{petab_applyInverseObservableTransformation}()},
\code{\link{petab_applyObservableTransformation}()},
\code{\link{petab_duplicateControls}()},
\code{\link{petab_measurementData}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{measurementData}
