% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getMeasurementParsMapping}
\alias{petab_getMeasurementParsMapping}
\title{Parse the parameters columns in measurementData}
\usage{
petab_getMeasurementParsMapping(
  pe,
  column = c("observableParameters", "noiseParameters")[1]
)
}
\arguments{
\item{column}{character(1L). one of c("observableParameters", "noiseParameters")}

\item{measurementData}{petab$measurementData. Uses only columns observableId,simulationConditionId and "column"}
}
\value{
data.table(condition, Innerpars... = Outerpars...) of local parameters.
Can be added to a gridlist with \code{\link[dMod:indiv_addLocalParsToGridList]{dMod::indiv_addLocalParsToGridList()}}
}
\description{
Parse the parameters columns in measurementData
}
\examples{
measurementData <- data.table(
observableId = c("obsE"),
simulationConditionId = c("C1", "C2"),
observableParameters = c("scale;offset", "1;offset"),
noiseParameters = "sigmaAbs"
)
experimentalCondition <- data.table(conditionId = c("C1", "C2", "C3"))
pe <- list(measurementData = measurementData, experimentalCondition = experimentalCondition)
petab_getMeasurementParsMapping(pe, "noiseParameters")
petab_getMeasurementParsMapping(pe, "observableParameters")
# suggested use: indiv_addLocalParsToGridlist ...
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
