% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getParameterBoundaries}
\alias{petab_getParameterBoundaries}
\title{Title}
\usage{
petab_getParameterBoundaries(pe, whichBoundary = c("upper", "lower")[1])
}
\arguments{
\item{pe}{petab}

\item{whichBoundary}{upper or lower}
}
\value{
c(name = value)
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
