% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{petab_getParameterFixedList}
\alias{petab_getParameterFixedList}
\title{Get parameter formulas from dMod trafo}
\usage{
petab_getParameterFixedList(trafoInfo)
}
\arguments{
\item{trafoInfo}{as output by petab_getTrafoInfo()}
}
\value{
DT with columns parameterId, parameterFormula and the attribute generalScale
}
\description{
Get parameter formulas from dMod trafo
}
\seealso{
Other Parameter wrangling: 
\code{\link{getEXgrid}()},
\code{\link{petab_getParameterFormulaList}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_getTrafoInfo}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()},
\code{\link{strsplit2}()}
}
\author{
Svenja Kemmer
}
\concept{Parameter wrangling}
