% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getParametersExperimentalCondition}
\alias{petab_getParametersExperimentalCondition}
\title{Get parameter names from pe$experimentalCondition}
\usage{
petab_getParametersExperimentalCondition(experimentalCondition)
}
\arguments{
\item{experimentalCondition}{}
}
\value{
character vector
}
\description{
Get parameter names from pe$experimentalCondition
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
