% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getParametersToEstimate}
\alias{petab_getParametersToEstimate}
\title{Title}
\usage{
petab_getParametersToEstimate(pe)
}
\arguments{
\item{pe}{}
}
\value{
character vector
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
