% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getParameterType}
\alias{petab_getParameterType}
\title{Get a mapping of parameter types}
\usage{
petab_getParameterType(pe)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}} (with parameters defined)}
}
\value{
data.table(parameterId, parameterType)
}
\description{
Get a mapping of parameter types
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
