% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getPars_estScale}
\alias{petab_getPars_estScale}
\title{Get parameters on outer scale}
\usage{
petab_getPars_estScale(pe)
}
\arguments{
\item{pe}{}
}
\description{
Get parameters on outer scale
}
\examples{
pe <- petab_exampleRead("01")
parsEst <- petab_getPars_estScale(pe)
}
\seealso{
Other Parameter wrangling: 
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Parameter wrangling}
