% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getPars_linScale}
\alias{petab_getPars_linScale}
\title{Get parameters on nominal scale}
\usage{
petab_getPars_linScale(pe)
}
\arguments{
\item{pe}{}
}
\value{

}
\description{
Get parameters on nominal scale
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
