% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{petab_getTrafoInfo}
\alias{petab_getTrafoInfo}
\title{Get parameter infos from dMod trafo}
\usage{
petab_getTrafoInfo(trafo)
}
\arguments{
\item{trafo}{as output by dMod::define()}
}
\value{
DT with columns parameterId, parameterFormula and trafoScale
}
\description{
Get parameter infos from dMod trafo
}
\seealso{
Other Parameter wrangling: 
\code{\link{getEXgrid}()},
\code{\link{petab_getParameterFixedList}()},
\code{\link{petab_getParameterFormulaList}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()},
\code{\link{strsplit2}()}
}
\author{
Svenja Kemmer
}
\concept{Parameter wrangling}
