% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_hash}
\alias{petab_hash}
\title{Hash a petab}
\usage{
petab_hash(pe)
}
\arguments{
\item{pe}{petab}
}
\value{
a digest
}
\description{
The difficulty is that data.tables have attributes specific to R sessions
}
\examples{
# Should evaluate to TRUE, else the petab has changed
petab_hash(petab_exampleRead("01","pe")) == "45281114c1d5f92ca00f9f2171379ba2"
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab helpers}
