% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_lint}
\alias{petab_lint}
\title{Own little linter until petab.py lint is stable}
\usage{
petab_lint(pe)
}
\arguments{
\item{petab}{}
}
\value{
list of errors
}
\description{
Own little linter until petab.py lint is stable
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
