% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_markDataPointsShiny_step1}
\alias{petab_markDataPointsShiny_step1}
\alias{petab_markDataPointsShiny_step2}
\title{Mark data points in a plot}
\usage{
petab_markDataPointsShiny_step1(plot, fileCSV)

petab_markDataPointsShiny_step2(fileCSV, filename = NULL, NFLAGtribble = 0)
}
\arguments{
\item{plot}{A plot from petab_plotData. Must have aeslist(customdata~datapointId) mapped}

\item{fileCSV}{filename for csv file}
}
\value{
Called for side-effect
}
\description{
\itemize{
\item Opens shiny app
\item Writes clicked points to a file
}
}
\examples{
# Plot data
pe <- petab_exampleRead(exampleName = "01", "pe")
pe$measurementData[,`:=`(datapointId = sprintf(paste0("\%s_\%0",nchar(as.character(.N)), "i"),  datasetId, 1:.N))]
plot <- petab_plotData(pe, aeslist = list(customdata=~datapointId))

# Mark outliers
fileCSV <- "~/wup.txt"
petab_markDataPointsShiny_step1(plot, fileCSV)
outliers <- petab_markDataPointsShiny_step2(fileCSV)

# Flag them in the original plot
pe$measurementData[,`:=`(outlier = FALSE)]
pe$measurementData[datapointId \%in\% outliers$datapointId, `:=`(outlier = TRUE)]
petab_plotData(pe, aeslist = list(customdata=~datapointId, shape = ~outlier, size = ~outlier))
}
\seealso{
Other plotting: 
\code{\link{pd_plotParsParallelLines2}()},
\code{\link{pd_plotParsParallelLines}()},
\code{\link{pd_plotProfile}()},
\code{\link{petab_plotData}()},
\code{\link{petab_plotHelpers_aeslist}()},
\code{\link{petab_plotHelpers_meanMeasurementsValue}()},
\code{\link{petab_plotHelpers_parameterOrder}()},
\code{\link{petab_plotHelpers_variableOrder}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotting}
