% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_mergeParameters}
\alias{petab_mergeParameters}
\title{Merge parameter values from one petab into another petab}
\usage{
petab_mergeParameters(
  pe1,
  pe2,
  mergeCols = c("nominalValue", "parameterScale", "estimate")
)
}
\arguments{
\item{pe1}{parameters from pe1}

\item{pe2}{parameters from pe2}

\item{mergeCols}{columns to merge into pe_pa1}
}
\value{
updated pe1
}
\description{
Merge parameter values from one petab into another petab
}
\seealso{
Other Parameter wrangling: 
\code{\link{getEXgrid}()},
\code{\link{petab_getParameterFormulas}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()},
\code{\link{strsplit2}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Parameter wrangling}
