% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_meta}
\alias{petab_meta}
\title{Additional Information}
\usage{
petab_meta(
  parameterFormulaInjection = NULL,
  variableOrder = NULL,
  metaInformation = NULL,
  ...
)
}
\arguments{
\item{parameterFormulaInjection}{data.table(parameterId, parameterFormula)}

\item{variableOrder}{vector of stateIds and observableIds in the order which you want for plotting}

\item{metaInformation}{ordered list potentially containing descriptions of columns, e.g. list(experimentalCondition = list(conditionId = list(patterns = "cellline_dose")), measurementData = list(time = list(unit = "hours"), measurement = list(unit = "value", lloq = 1))) see example of petab_joinDCO}
}
\value{
named list of supplied arguments
}
\description{
Additional Information
}
