% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_metaInformation_getPatterns}
\alias{petab_metaInformation_getPatterns}
\title{Get "patterns" from metaInformation}
\usage{
petab_metaInformation_getPatterns(mi)
}
\arguments{
\item{pe}{pe with pe$meta$metaInformation}
}
\value{
list(column_name = "pattern")
}
\description{
Get "patterns" from metaInformation
}
\examples{
# In two tables
mi <- yaml::read_yaml(text = "measurementData:\n  measurement:\n    unit: absolute_conc\n    lloq: ~\n  time:\n    unit: hours\n  replicateId:\n    pattern: experiment_replicate_technicalReplicate\nexperimentalCondition:\n  conditionId:\n    pattern: cellline_TGFb_Statin\n  TGFb:\n    unit: ng/ul\n  Statin:\n    unit: ug/ml\n")
petab_metaInformation_getPatterns(mi)

# In one table
mi <- yaml::read_yaml(text = "measurementData:\n  measurement:\n    unit: absolute_conc\n    lloq: ~\n  time:\n    unit: hours\n  experimentalCondition:\n  conditionId:\n    pattern: cellline_TGFb_Statin\n  TGFb:\n    unit: ng/ul\n  Statin:\n    unit: ug/ml\n")
petab_metaInformation_getPatterns(mi)

# No pattern
mi <- yaml::read_yaml(text = "measurementData:\n  measurement:\n    unit: absolute_conc")
petab_metaInformation_getPatterns(mi)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{metaInformation}
