% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_model}
\alias{petab_model}
\title{PEtab structural model without sbml}
\usage{
petab_model(
  equationList,
  events = NA,
  parInfo = getParInfo(equationList),
  speciesInfo = getSpeciesInfo(equationList),
  ...
)
}
\arguments{
\item{equationList}{eqnlist}

\item{events}{eventlist}

\item{...}{not used, but could be used in the future for imitating assignment rules etc}

\item{parameterFormulaInjection}{}
}
\value{
list
}
\description{
PEtab structural model without sbml
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
