% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_mutateDCO}
\alias{petab_mutateDCO}
\title{Consistently mutate the dco}
\usage{
petab_mutateDCO(pe, i, j)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}}}

\item{i}{expression to subset the dco, can be missing}

\item{j}{Must be a call of the form \verb{:=}(hsdhfgjs). If missing, pe is subsetted by i}
}
\value{
pe with modified DCO
}
\description{
If j is supplied, the full dco is always returned and no subsetting is done!
}
\examples{
# todo!!
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
