% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_overview}
\alias{petab_overview}
\title{Print some overview stuff}
\usage{
petab_overview(
  pe,
  FLAGcolumns = TRUE,
  FLAGmetaInformation = TRUE,
  FLAGobsPerCond = TRUE,
  ...
)
}
\arguments{
\item{pe}{pe}

\item{FLAGcolumns, FLAGmetaInformation, FLAGobsPerCond}{include this and that info?}

\item{...}{arguments going to \code{\link[=petab_overviewObsPerCond]{petab_overviewObsPerCond()}}}
}
\value{
Prints some stuff to console
}
\description{
Print some overview stuff
}
\examples{
petab_overview(petab_exampleRead("01", "pe"))
}
\seealso{
Other Overview Tables: 
\code{\link{petab_overviewDCONames}()},
\code{\link{petab_overviewObsPerCond}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Overview Tables}
