% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_overviewDCONames}
\alias{petab_overviewDCONames}
\title{Print available names in the DCO}
\usage{
petab_overviewDCONames(pe)
}
\arguments{
\item{pe}{}
}
\value{
prints list of available columns
}
\description{
Print available names in the DCO
}
\seealso{
Other Overview Tables: 
\code{\link{petab_overviewObsPerCond}()},
\code{\link{petab_overview}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Overview Tables}
