% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_overviewObsPerCond}
\alias{petab_overviewObsPerCond}
\title{Generate overview table which observables are in which condition}
\usage{
petab_overviewObsPerCond(pe, Ntruncate = 1000, ...)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}} object}

\item{Ntruncate}{truncate pasted observables at this many characters}

\item{...}{}
}
\value{

}
\description{
Generate overview table which observables are in which condition
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
