% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_parameterFormulaInjection}
\alias{petab_parameterFormulaInjection}
\title{Create a parameterFormulaInjection}
\usage{
petab_parameterFormulaInjection(parameterId, parameterFormula)
}
\arguments{
\item{parameterId}{e.g. "kprodS"}

\item{parameterFormula}{"kdegS * S"}
}
\value{
data.table(parameterId, parameterFormula)
}
\description{
See example-02
}
\details{
Inject a trafo between pEst and x
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
