% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_collectCriteria}
\alias{petab_select_collectCriteria}
\title{Title}
\usage{
petab_select_collectCriteria(AIC, BIC)
}
\arguments{
\item{AIC, BIC}{numeric(1L), the respective criteria}
}
\value{
list(AIC, BIC)
}
\description{
Title
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_createReportYaml}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
