% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_collectCriteriaFromPD}
\alias{petab_select_collectCriteriaFromPD}
\title{Title}
\usage{
petab_select_collectCriteriaFromPD(pd)
}
\arguments{
\item{pd}{a pd}
}
\value{
list(AIC = numeric(1L), BIC = numeric(1L))
}
\description{
Title
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_createReportYaml}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
