% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_collectEstimatedParametersFromPD}
\alias{petab_select_collectEstimatedParametersFromPD}
\title{Collect list of estimated parameters}
\usage{
petab_select_collectEstimatedParametersFromPD(pd)
}
\arguments{
\item{pd}{a pd}
}
\value{
list(parameter = value)
}
\description{
Chooses based on availability
\enumerate{
\item If mstrust is available, take from mstrust
\item else take from "base" (which might be a single fit) CAREFUL: base might be worse than base_obsParsFitted, as base is not necessarily fitted
\item else take from "base_obsParsFitted"
\item else take from pd$pars (which might be unfitted)
}
}
\examples{
pd <- petab_exampleRead("01", "pd")
petab_select_collectEstimatedParametersFromPD(pd)
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_createReportYaml}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
