% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_collectParametersFromPD}
\alias{petab_select_collectParametersFromPD}
\title{Collect Parameters from PD}
\usage{
petab_select_collectParametersFromPD(pd)
}
\arguments{
\item{pd}{a pd}
}
\value{
named list(fixedPar = value, estimatedPar = "estimate")
}
\description{
Collect Parameters from PD
}
\examples{
pd <- petab_exampleRead("01", "pd")
petab_select_collectParametersFromPD(pd)
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_createReportYaml}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
