% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_createReportYaml}
\alias{petab_select_createReportYaml}
\title{Collect all information for the reportYaml}
\usage{
petab_select_createReportYaml(
  model_id,
  petab_yaml,
  sbml,
  parameters,
  estimated_parameters,
  criteria
)
}
\arguments{
\item{model_id}{string}

\item{petab_yaml}{filename of yaml}

\item{sbml}{filename of sbml}

\item{parameters}{named list of numeric entries for fixed parameters and "estimate" for estimated parameters}

\item{estimated_parameters}{named list of numeric parameters}

\item{criteria}{list(LL, AIC, BIC)}
}
\value{
simply a list of the supplied objects
}
\description{
Collect all information for the reportYaml
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
