% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_createReportYamlFromPD}
\alias{petab_select_createReportYamlFromPD}
\title{Create the reportYaml from a PD}
\usage{
petab_select_createReportYamlFromPD(pd, FLAGwriteYaml = TRUE)
}
\arguments{
\item{pd}{a pd}

\item{FLAGwriteYaml}{TRUE: Writes reportYaml into the original petab folder}
}
\value{
reportYaml as list, see \link{petab_select_createReportYaml}
}
\description{
Create the reportYaml from a PD
}
\examples{
# create reportYaml
pd <- petab_exampleRead("01", "pd")
reportYaml <- petab_select_createReportYamlFromPD(pd, FLAGwriteYaml = FALSE)

# write it to disk and look at it
tf <- tempfile(fileext = ".yaml")
petab_select_writeReportYaml(reportYaml, tf)
file.edit(tf)
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYaml}()},
\code{\link{petab_select_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
