% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petab_select_writeReportYaml}
\alias{petab_select_writeReportYaml}
\title{Title}
\usage{
petab_select_writeReportYaml(
  reportYaml,
  filename = petab_files(reportYaml$model_id)["reportYaml"]
)
}
\arguments{
\item{reportYaml}{reportYaml as list, see \link{petab_select_createReportYaml}}

\item{filename}{filename to write to}
}
\value{
NULL, called for side-effect
}
\description{
Title
}
\seealso{
Other petab select: 
\code{\link{petab_select_collectCriteriaFromPD}()},
\code{\link{petab_select_collectCriteria}()},
\code{\link{petab_select_collectEstimatedParametersFromPD}()},
\code{\link{petab_select_collectParametersFromPD}()},
\code{\link{petab_select_createReportYamlFromPD}()},
\code{\link{petab_select_createReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
