% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_transformPars_est2Lin}
\alias{petab_transformPars_est2Lin}
\title{Title}
\usage{
petab_transformPars_est2Lin(pe, parsEst)
}
\arguments{
\item{parsEst}{}
}
\value{
vector
}
\description{
Title
}
\examples{
pe <- petab_exampleRead("01")
parsEst <- petab_getPars_estScale(pe)
parsLin <- petab_getPars_linScale(pe)
parsLin <- parsLin[sample(1:length(parsLin), 5)] + rnorm(5)
}
\seealso{
Other Parameter wrangling: 
\code{\link{getEXgrid}()},
\code{\link{petab_getParameterFormulas}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_lin2Est}()},
\code{\link{strsplit2}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Parameter wrangling}
