% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_unjoinDCO}
\alias{petab_unjoinDCO}
\title{Unjoin DataConditonObs}
\usage{
petab_unjoinDCO(DCO, pe = NULL)
}
\arguments{
\item{DCO}{A dco: a DataConditionObs data.table, from petab_joinDCO}

\item{pe}{\code{\link[=petab]{petab()}} object}
}
\value{
pe with measurementData,experimentalCondition,Observables replaced by the columns created from DCO
}
\description{
Unjoin DataConditonObs
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
