% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_compareModels}
\alias{petabSelect_compareModels}
\title{Compare models}
\usage{
petabSelect_compareModels(reportYamlFilenames, sort_by = "BIC")
}
\arguments{
\item{reportYamlFilenames}{}

\item{sort_by}{column to sort models}
}
\value{
data.table
}
\description{
Compare models
}
\examples{
# TODO
}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
