% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_LRT}
\alias{petabSelect_LRT}
\title{Likelihood ratio test}
\usage{
petabSelect_LRT(comparisonTable, alpha = 0.05)
}
\arguments{
\item{comparisonTable}{output from \code{\link[=petabSelect_compareModels]{petabSelect_compareModels()}} with two nested models}

\item{alpha}{Significance level of statistical test}
}
\value{
Prints a message, returns the printed table invisbly
}
\description{
Likelihood ratio test
}
\examples{
comparisonTable <- data.table(tibble::tribble(
~model_id, ~nEstimatedPars   , ~`-2LL`, 
"Model_full"   ,            2, 5,
"Model_reduced",            1, 9))
}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
