% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_readReportYaml}
\alias{petabSelect_readReportYaml}
\title{Read report yaml content}
\usage{
petabSelect_readReportYaml(filename)
}
\arguments{
\item{filename}{file path}
}
\value{
the yaml contents as list
}
\description{
Read report yaml content
}
\examples{
pd <- petab_exampleRead("04", "pd")
reportYaml <- pd_petabSelect_reportYaml(pd,F)
tf <- tempfile(fileext = ".yaml")
petabSelect_writeReportYaml(reportYaml, tf)
reportYaml_fromFile <- petabSelect_readReportYaml(tf)

}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_LRT}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
