% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_reportYaml}
\alias{petabSelect_reportYaml}
\title{Collect all information for the reportYaml}
\usage{
petabSelect_reportYaml(
  model_id,
  petab_yaml,
  sbml,
  parameters,
  estimated_parameters,
  criteria
)
}
\arguments{
\item{model_id}{string}

\item{petab_yaml}{filename of yaml}

\item{sbml}{filename of sbml}

\item{parameters}{named list of numeric entries for fixed parameters and "estimate" for estimated parameters}

\item{estimated_parameters}{named list of numeric parameters}

\item{criteria}{list(AIC, BIC, AICc)}
}
\value{
simply a list of the supplied objects
}
\description{
Collect all information for the reportYaml
}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
