% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_searchYamls}
\alias{petabSelect_searchYamls}
\title{List yaml files in subdirectories based on a pattern}
\usage{
petabSelect_searchYamls(path, pattern)
}
\arguments{
\item{path}{}

\item{pattern}{}
}
\description{
List yaml files in subdirectories based on a pattern
}
