% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_writeReportYaml}
\alias{petabSelect_writeReportYaml}
\title{Title}
\usage{
petabSelect_writeReportYaml(
  reportYaml,
  filename = petab_files(reportYaml$model_id)["reportYaml"]
)
}
\arguments{
\item{reportYaml}{reportYaml as list, see \link{petabSelect_reportYaml}}

\item{filename}{filename to write to}
}
\value{
NULL, called for side-effect
}
\description{
Title
}
\seealso{
Other petab select: 
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_collectReportYamlCriteria}()},
\code{\link{petabSelect_reportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
