% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petabYaml}
\alias{petabYaml}
\title{Gather all filenames in the list structure of a petab.yaml}
\usage{
petabYaml(
  parameter_file = NULL,
  condition_files = NULL,
  measurement_files = NULL,
  observable_files = NULL,
  sbml_files = NULL,
  visualization_files = NULL,
  meta_file = NULL,
  metaInformation_file = NULL,
  filename = NULL
)
}
\arguments{
\item{condition_files, measurement_files, observable_files, sbml_files, visualization_files, meta_file, metaInformation_file, parameter_file}{file.path of length 1 or NULL}

\item{filename}{file.path to write yaml to}
}
\value{
list of these objects, with sublist "problems"
}
\description{
Gather all filenames in the list structure of a petab.yaml
}
\examples{
petabYaml(parameter_file = "wup", 
condition_files = NULL, 
measurement_files = "bla", 
observable_files = NULL, 
sbml_files = NULL, 
visualization_files = NULL,
meta_file = NULL, 
metaInformation_file = "bam", filename = "bla.yaml")
}
\seealso{
Other petab_files: 
\code{\link{petab_files_default}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab_files}
