% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{plotParameterComparison}
\alias{plotParameterComparison}
\title{Compare two parameter vectors by plotting their difference as bar plot}
\usage{
plotParameterComparison(
  v1,
  v2,
  pe,
  filename,
  FLAGexcludeObsPars = TRUE,
  FLAGestimatedOnly = TRUE
)
}
\arguments{
\item{v1, v2}{Parameters on estScale}

\item{pe}{Petab to infer parameterType and estimated parameters}

\item{filename}{for plotting}
}
\value{
ggplot
}
\description{
Compare two parameter vectors by plotting their difference as bar plot
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
