% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{qNLME_getParameterFormulaInjection}
\alias{qNLME_getParameterFormulaInjection}
\title{Create qNLME Trafo, depending on parameterScale}
\usage{
qNLME_getParameterFormulaInjection(pe, parameterId_base)
}
\arguments{
\item{pe}{}

\item{parameterId_base}{vector of parameterId which should be qNLME'd.}
}
\value{
data.table(parameterId, parameterFormula, trafoType = "qNLME")
}
\description{
Trafo injection
}
\details{
\itemize{
\item For logpars: par * qNLME_par
\item For linpars: par + qNLME_par
}
}
\examples{
pe <- petab_exampleRead("05")
qNLME_getParameterFormulaInjection(pe, c("kcat", "E"))
}
\seealso{
Other petab qNLME: 
\code{\link{pe_qNLME_updateParameterFormulaInjection}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab qNLME}
