% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{readPetab}
\alias{readPetab}
\title{Read PEtab files}
\usage{
readPetab(filename)
}
\arguments{
\item{path}{}
}
\value{

}
\description{
Read PEtab files
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
