% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re-exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{n_pages}
\alias{facet_grid_paginate}
\alias{facet_wrap_paginate}
\alias{\%>\%}
\alias{tidyverse_logo}
\title{Reexports}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggforce}{\code{\link[ggforce]{facet_grid_paginate}}, \code{\link[ggforce]{facet_wrap_paginate}}, \code{\link[ggforce]{n_pages}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tidyverse}{\code{\link[tidyverse]{tidyverse_logo}}}
}}

