% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneParameter}
\alias{sbml_addOneParameter}
\title{Title}
\usage{
sbml_addOneParameter(model, parName, parValue, parUnit)
}
\arguments{
\item{model}{}

\item{speciesName}{}

\item{compName}{}

\item{speciesInitialAmount}{}
}
\description{
Title
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
