% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneReaction}
\alias{sbml_addOneReaction}
\title{Title}
\usage{
sbml_addOneReaction(
  model,
  reactionName,
  reversible = 1,
  reactants,
  products,
  equation,
  parName,
  parValue,
  parUnit,
  modifiers
)
}
\arguments{
\item{model}{}

\item{reactionName}{}

\item{reversible}{}

\item{reactants}{}

\item{products}{}

\item{equation}{}

\item{parName}{}

\item{parValue}{}

\item{parUnit}{}
}
\description{
Title
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameter}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
