% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_addOneSpeciesSym}
\alias{sbml_addOneSpeciesSym}
\title{Title}
\usage{
sbml_addOneSpeciesSym(model, speciesName, compName, initialAmount)
}
\arguments{
\item{model}{}

\item{speciesName}{}

\item{compName}{}

\item{speciesInitialAmount}{}
}
\value{

}
\description{
Title
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfoSimple}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameterNum}()},
\code{\link{sbml_addOneParameterSym}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpeciesNum}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Svenja Kemmer
}
\concept{SBML export}
