% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_initialize}
\alias{sbml_initialize}
\title{Initialize the sbml document}
\usage{
sbml_initialize(modelname = "EnzymaticReaction", sbmlDoc)
}
\arguments{
\item{modelname}{String}

\item{sbmlDoc}{return value from a call like \code{SBMLDocument(level = 2, version = 4)}}
}
\value{
a libsbml model representation
}
\description{
Initialize the sbml document
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfoSimple}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameterNum}()},
\code{\link{sbml_addOneParameterSym}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpeciesNum}()},
\code{\link{sbml_addOneSpeciesSym}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
