% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sbmlImport_getFunctionInfo}
\alias{sbmlImport_getFunctionInfo}
\title{Get function info}
\usage{
sbmlImport_getFunctionInfo(fun)
}
\arguments{
\item{fun}{some sbml-thingy representation of a function}
}
\value{
list with infos about the funciton
}
\description{
function adapted from Frank
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML import
importFrom libSBML formulaToL3String}
