% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sbmlImport_getInitialsFromSBML}
\alias{sbmlImport_getInitialsFromSBML}
\title{Import initials from SBML.}
\usage{
sbmlImport_getInitialsFromSBML(modelfile)
}
\arguments{
\item{model}{SBML file as .xml}
}
\value{
Named vector of initials.
}
\description{
This function imports initial values or equations describing the same from SBML and writes them in a named vector.
}
\author{
Marcus Rosenblatt, Svenja Kemmer and Frank Bergmann
importFrom libSBML readSBML formulaToL3String
}
