% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sbmlImport_getParametersFromSBML}
\alias{sbmlImport_getParametersFromSBML}
\title{Extract parameters from SBML}
\usage{
sbmlImport_getParametersFromSBML(modelfile)
}
\arguments{
\item{model}{SBML file as .xml}
}
\value{
List of parameters vectors: pars_num and pars_sym
}
\description{
This function imports parameter values or equations describing the same from SBML and writes them in a named vector.
}
\author{
Marcus Rosenblatt, Svenja Kemmer and Frank Bergmann
importFrom libSBML readSBML formulaToL3String
}
