% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{scale_color_cf}
\alias{scale_color_cf}
\title{scalecolorcf
Copied from dMod}
\usage{
scale_color_cf(...)
}
\arguments{
\item{...}{see scale_color_manual}
}
\value{
added to plots
}
\description{
scalecolorcf
Copied from dMod
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
