% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{scale_inverse}
\alias{scale_inverse}
\title{Title}
\usage{
scale_inverse(parameterScale)
}
\arguments{
\item{parameterScale}{vector with entries "lin", "log", "log10"}
}
\value{

}
\description{
Title
}
\seealso{
Other scales: 
\code{\link{lin}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{scales}
