% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{smblImport_replaceFunctions}
\alias{smblImport_replaceFunctions}
\title{Replace function calls by their mathematical expressions}
\usage{
smblImport_replaceFunctions(m, fun, rates)
}
\arguments{
\item{m}{sbml model}

\item{fun}{an integer with a function id (badly refactored)}

\item{rates}{reaction rate formulas}
}
\value{
modified rates
}
\description{
Replace function calls by their mathematical expressions
}
\seealso{
Other SBML import: 
\code{\link{sbmlImport_getCompartmentsFromSBML}()},
\code{\link{sbmlImport_getParametersFromSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML import}
