% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{strsplit2}
\alias{strsplit2}
\title{strsplit with additional position argument}
\usage{
strsplit2(x, split, type = "remove", perl = FALSE, ...)
}
\arguments{
\item{x}{character to split}

\item{split}{pattern as split criterion}

\item{type}{position of split c("remove", "before", "after")}

\item{...}{}
}
\value{
character vector with split x
}
\description{
strsplit with additional position argument
}
\seealso{
Other Parameter wrangling: 
\code{\link{getEXgrid}()},
\code{\link{petab_getParameterFormulas}()},
\code{\link{petab_getPars_estScale}()},
\code{\link{petab_getPars_linScale}()},
\code{\link{petab_mergeParameters}()},
\code{\link{petab_parameters_mergeParameters}()},
\code{\link{petab_setPars_estScale}()},
\code{\link{petab_setPars_linScale}()},
\code{\link{petab_transformPars_est2Lin}()},
\code{\link{petab_transformPars_lin2Est}()}
}
\author{
Svenja Kemmer
}
\concept{Parameter wrangling}
