% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copied_from_conveniencefunctions.R
\name{write_clusterTimeStamp}
\alias{write_clusterTimeStamp}
\title{Write the time stamp when you logged into the cluster}
\usage{
write_clusterTimeStamp()
}
\description{
Write the time stamp when you logged into the cluster
}
\seealso{
Other Cluster login helpers: 
\code{\link{check_clusterTimeStamp}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster login helpers}
