% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePE.R
\name{writePE}
\alias{writePE}
\title{Write PE from dMod objects}
\usage{
writePE(
  modelname = "mymodel",
  ODEmodel = reactions,
  obs_fun = observables,
  errormodel = errors,
  data = pred,
  parameters = bestfit,
  conditions = condition.grid,
  est_grid = est.grid,
  fixed_grid = fixed.grid,
  parUnit = "per_minute"
)
}
\arguments{
\item{modelname}{}

\item{ODEmodel}{}

\item{obs_fun}{}

\item{errormodel}{}

\item{data}{}

\item{parameters}{}

\item{conditions}{}

\item{parUnit}{}
}
\value{
pe list
}
\description{
Write PE from dMod objects
}
\author{
Svenja Kemmer
}
