% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getParInfo}
\alias{getParInfo}
\title{Extract parInfo data.table from equationlist}
\usage{
getParInfo(
  equationList,
  eventList = NULL,
  parameterFormulaList = NULL,
  unit = "identity"
)
}
\arguments{
\item{unit}{}
}
\value{
data.table(parName,parValue,parUnit)
}
\description{
Extract parInfo data.table from equationlist
}
\examples{
library(dMod)
example(eqnlist)
getParInfo(f)
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameter}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
