% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{getReactionInfo}
\alias{getReactionInfo}
\title{Title}
\usage{
getReactionInfo(equationList, parInfo = getParInfo(equationList))
}
\arguments{
\item{el}{}

\item{p}{}
}
\value{
list of reactions
}
\description{
Title
}
\examples{

# simple
el <- NULL
el <- addReaction(el, from = "E + S", to = "ES", rate = "(kon)*E*S",
                  description = "production of complex")
el <- addReaction(el, from = "ES", to = "E + S", rate = "koff*ES",
                  description = "decay of complex")
el <- addReaction(el, from = "ES", to = "E + P", rate = "kcat*ES",
                  description = "production of product")
el <- eqnlist_addDefaultCompartment(el, "cytoplasm") # Need compartment information for SBML
equationList <- el
parInfo = getParInfo(equationList)

# complex stoichiometries, modifier in production of complex
el <- NULL
el <- addReaction(el, from = "2*E + S", to = "ES", rate = "(kon)*E*S * 1/(1+kinh*P)",
                  description = "production of complex")
el <- addReaction(el, from = "ES", to = "2*E + S", rate = "koff*ES",
                  description = "decay of complex")
el <- addReaction(el, from = "ES", to = "E + P", rate = "kcat*ES",
                  description = "production of product")
el <- eqnlist_addDefaultCompartment(el, "cytoplasm") # Need compartment information for SBML
equationList <- el
parInfo = getParInfo(equationList)

getReactionInfo(equationList,parInfo)
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameterNum}()},
\code{\link{sbml_addOneParameterSym}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpeciesNum}()},
\code{\link{sbml_addOneSpeciesSym}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_exportEquationList}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)

Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
