% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{initPd}
\alias{initPd}
\title{Title}
\usage{
initPd(filenameParts, ...)

initPd(filenameParts, ...)
}
\arguments{
\item{filenameParts}{}

\item{...}{}

\item{pe}{}

\item{dModAtoms}{}

\item{obj}{}

\item{prd}{}

\item{times}{}

\item{pars}{}

\item{parameterTables}{}

\item{L1}{}

\item{pdInfo}{}
}
\value{



}
\description{
Title
}
