% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_cluster_L1}
\alias{pd_cluster_L1}
\title{Fit model on cluster}
\usage{
pd_cluster_L1(
  pd,
  .outputFolder,
  n_nodes = 6,
  lambdas = 10^(seq(log10(1e-04), log10(100), length.out = n_nodes * 16 - 1)),
  identifier = "L1",
  FLAGforcePurge = FALSE
)
}
\arguments{
\item{pd}{}

\item{.outputFolder}{}

\item{n_nodes}{}

\item{n_startsPerNode}{}

\item{id}{}

\item{type}{}
}
\value{
Characters
}
\description{
Fit model on cluster
}
\seealso{
Other Cluster: 
\code{\link{clusterStatusMessage}()},
\code{\link{pd_cluster_L1_fitUnbiasedKNECHT}()},
\code{\link{pd_cluster_L1_fitUnbiased}()},
\code{\link{pd_cluster_mstrust}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Cluster}
