% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fit}
\alias{pd_fit}
\title{Run a fit only for observation parameters}
\usage{
pd_fit(pd, iterlim = 1000, printIter = TRUE, FLAGoverwrite = FALSE, ...)
}
\arguments{
\item{iterlim}{}

\item{printIter}{}

\item{FLAGoverwrite}{}

\item{...}{other arguments going to trust}
}
\value{

}
\description{
Run a fit only for observation parameters
}
\seealso{
Other pd: 
\code{\link{pd_mstrust}()},
\code{\link{pd_rdsfile}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd}
\concept{pd fittting}
