% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fitMstrust}
\alias{pd_fitMstrust}
\title{Run a mstrust locally}
\usage{
pd_fitMstrust(
  pd,
  fits = 20,
  iterlim = 1000,
  FLAGoverwrite = FALSE,
  FLAGjobPurgedFile = TRUE,
  ...
)
}
\arguments{
\item{fits}{number of fits, see \code{\link[dMod:mstrust]{dMod::mstrust()}}}

\item{iterlim}{}

\item{FLAGoverwrite}{}

\item{FLAGjobPurgedFile}{add a jobPurged-file as if the mstrust was run by a pd_cluster function}

\item{...}{other arguments going to \code{\link[dMod:mstrust]{dMod::mstrust()}}}

\item{printIter}{}
}
\value{

}
\description{
Run a mstrust locally
}
\seealso{
Other pd: 
\code{\link{pd_fit}()},
\code{\link{pd_mstrust}()},
\code{\link{pd_rdsfile}()}

Other pd fittting: 
\code{\link{pd_fit}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd}
\concept{pd fittting}
