% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fitMstrust}
\alias{pd_fitMstrust}
\title{Run a mstrust locally}
\usage{
pd_fitMstrust(
  pd,
  fits = 20,
  iterlim = 1000,
  printIter = TRUE,
  FLAGoverwrite = FALSE,
  ...
)
}
\arguments{
\item{iterlim}{}

\item{printIter}{}

\item{FLAGoverwrite}{}

\item{...}{other arguments going to dMod::mstrust}
}
\value{

}
\description{
Run a mstrust locally
}
\seealso{
Other pd: 
\code{\link{pd_fit}()},
\code{\link{pd_mstrust}()},
\code{\link{pd_rdsfile}()}

Other pd fittting: 
\code{\link{pd_fit}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd}
\concept{pd fittting}
