% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fitObsPars}
\alias{pd_fitObsPars}
\title{Run a fit only for observation parameters}
\usage{
pd_fitObsPars(pd, NFLAGsavePd = 3)
}
\arguments{
\item{NFLAGsavePd}{as in pd_importIndiv: 0: don't save, 1: save, but redo everything, 3: if it was done before and pd didn't change, just return the pd}
}
\value{

}
\description{
Run a fit only for observation parameters
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
