% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_fitObsPars}
\alias{pd_fitObsPars}
\title{Run a fit only for observation parameters}
\usage{
pd_fitObsPars(pd, FLAGsavePd = TRUE)
}
\arguments{
\item{FLAGsavePd}{}
}
\value{

}
\description{
Run a fit only for observation parameters
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
