% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_mstrust}
\alias{pd_mstrust}
\title{Run mstrust}
\usage{
pd_mstrust(pd, NFLAGsavePd = T, iterlim = 1000, nfits = 5, id)
}
\arguments{
\item{NFLAGsavePd}{3 as an option doesn't make sense hre}
}
\value{

}
\description{
Run mstrust
}
\seealso{
Other pd: 
\code{\link{pd_fit}()}

Other pd fitting: 
\code{\link{pd_fitObsPars}()}
}
\author{
svenja kemmer
}
\concept{pd}
\concept{pd fitting}
