% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_objtimes}
\alias{pd_objtimes}
\title{Wrapper around predtimes}
\usage{
pd_objtimes(pd, N = 100)
}
\arguments{
\item{N}{}
}
\value{

}
\description{
Wrapper around predtimes
}
\seealso{
Other pd helpers: 
\code{\link{pd_predtimes}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pd helpers}
