% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_parf_collectProfile}
\alias{pd_parf_collectProfile}
\title{Title}
\usage{
pd_parf_collectProfile(
  pd,
  rows = c("profile_endpoints", "optimum"),
  parameters = NULL
)
}
\arguments{
\item{parameters}{}
}
\value{
\code{\link[dMod:parframe]{dMod::parframe()}}
}
\description{
Title
}
\seealso{
Other parframe handling: 
\code{\link{pd_parf_collectL1}()},
\code{\link{pd_parf_collectMstrust}()},
\code{\link{pd_parf_collectPars}()},
\code{\link{pd_parf_collect}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{parframe handling}
