% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_plot_compareParameters}
\alias{pd_plot_compareParameters}
\title{Title}
\usage{
pd_plot_compareParameters(
  pd,
  parf,
  nrow = 1,
  ncol = 3,
  scales = "free",
  page = 1,
  ...
)
}
\arguments{
\item{parf}{}
}
\value{

}
\description{
Title
}
