% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_plotProfile}
\alias{pd_plotProfile}
\title{Title}
\usage{
pd_plotProfile(pd, ggCallback = NULL, nrow = 3, ncol = 4, ...)
}
\arguments{
\item{pd}{}

\item{filename}{}

\item{base_size}{}
}
\value{

}
\description{
Title
}
\seealso{
Other plotting: 
\code{\link{pd_plotParsParallelLines}()},
\code{\link{pd_plotProfilePaths}()},
\code{\link{petab_plotData}()},
\code{\link{petab_plotHelpers_aeslist}()},
\code{\link{petab_plotHelpers_meanMeasurementsValue}()},
\code{\link{petab_plotHelpers_parameterOrder}()},
\code{\link{petab_plotHelpers_variableOrder}()},
\code{\link{stepMax}}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotting}
