% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_predictAndPlot2}
\alias{pd_predictAndPlot2}
\title{Title}
\usage{
pd_predictAndPlot2(
  pd,
  pe = pd$pe,
  i,
  opt.base = pd_parf_opt.base(),
  opt.mstrust = pd_parf_opt.mstrust(),
  opt.profile = pd_parf_opt.mstrust(FALSE),
  NFLAGsubsetType = c(none = 0, strict = 1, keepInternal = 2)[2],
  FLAGsummarizeProfilePredictions = TRUE,
  FLAGmeanLine = TRUE,
  aeslist = petab_plotHelpers_aeslist(),
  ggCallback = list(facet_wrap_paginate(~observableId, nrow = 4, ncol = 4, scales =
    "free"), scale_y_continuous(n.breaks = 5)),
  filename = NULL,
  FLAGfuture = TRUE,
  width = 29.7,
  height = 21,
  scale = 1,
  units = "cm"
)
}
\arguments{
\item{units}{}
}
\value{

}
\description{
Title
}
