% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_predictAndPlot2}
\alias{pd_predictAndPlot2}
\title{Title}
\usage{
pd_predictAndPlot2(
  pd,
  pe = pd$pe,
  i,
  opt.base = pd_parf_opt.base(),
  opt.mstrust = pd_parf_opt.mstrust(),
  opt.profile = pd_parf_opt.profile(FALSE),
  opt.L1 = pd_parf_opt.L1(FALSE),
  NFLAGsubsetType = c(none = 0, strict = 1, keepInternal = 2)[2],
  FLAGsummarizeProfilePredictions = TRUE,
  FLAGmeanLine = FALSE,
  aeslist = petab_plotHelpers_aeslist(),
  ggCallback = list(facet_wrap_paginate(~observableId, nrow = 4, ncol = 4, scales =
    "free"), scale_y_continuous(n.breaks = 5)),
  opt.sim = list(Ntimes_gt5ParSetIds = 60),
  opt.gg = list(ribbonAlpha = 0.2),
  filename = NULL,
  FLAGfuture = TRUE,
  width = 29.7,
  height = 21,
  scale = 1,
  units = "cm",
  FLAGreturnPlotData = FALSE
)
}
\arguments{
\item{units}{}
}
\value{

}
\description{
Title
}
\examples{
# pd <- petab_exampleRead("02", "pd")
pe = pd$pe
opt.base = pd_parf_opt.base(FALSE,)
opt.mstrust = pd_parf_opt.mstrust(FALSE, fitrankRange = 1:2)
opt.profile <- pd_parf_opt.profile(T, parameters = c("geneSerpine1_act2", "geneSerpine1_inh1", "geneSerpine1_inh2", "geneSerpine1_inh3", "pRec_degind", "S_phos"))
NFLAGsubsetType = 0
FLAGsummarizeProfilePredictions = TRUE
FLAGmeanLine = FALSE
aeslist = petab_plotHelpers_aeslist()
ggCallback = list(facet_wrap_paginate(~observableId, nrow = 4, ncol = 4, scales = "free"), scale_y_continuous(n.breaks = 5))
opt.gg = list(ribbonAlpha = 0.2)
filename = NULL
FLAGfuture = TRUE
width = 29.7
height = 21
scale = 1
units = "cm"
}
\seealso{
Other Plotting: 
\code{\link{pd_plot}()},
\code{\link{pd_predictAndPlot}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Plotting}
