% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_tests}
\alias{pd_tests}
\title{Some tests}
\usage{
pd_tests(
  pd,
  page = 1,
  cn = 1,
  whichTests = c(plot = 1, objData = 2, derivs = 3, `Rprof obj data` = 4)[2]
)
}
\arguments{
\item{cn}{}
}
\value{

}
\description{
Some tests
}
