% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_tests}
\alias{pd_tests}
\title{Some tests}
\usage{
pd_tests(
  pd,
  page = 1,
  cn = 1,
  whichTests = c("01-plot", "02-objData", "03-derivs")[2]
)
}
\arguments{
\item{cn}{}
}
\value{

}
\description{
Some tests
}
