% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_updateData}
\alias{pd_updateData}
\title{Title}
\usage{
pd_updateData(pd, measurementDataNew, .outputFolder, pdCopyName = "pd_1000")
}
\arguments{
\item{pd}{a pd}

\item{measurementDataNew}{new measurement data}

\item{.outputFolder}{}

\item{pdCopyName}{}
}
\value{

}
\description{
Title
}
\examples{

pd <- petab_exampleRead("04", "pd")
pd$filenameParts$.projectFolder <- dirname(pd$filenameParts$.compiledFolder)
measurementDataNew <- copy(pd$pe$measurementData)[,`:=`(measurement = exp(log(measurement) + rnorm(.N, 0, .05)))]
}
\seealso{
Other simulation of data: 
\code{\link{subsetPredictionToData}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{simulation of data}
