% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_updateEstPars}
\alias{pd_updateEstPars}
\title{Update parameters}
\usage{
pd_updateEstPars(pd, parsEst, FLAGupdatePE = TRUE, FLAGsavePd = FALSE)
}
\arguments{
\item{FLAGupdatePE}{}
}
\value{

}
\description{
Update parameters
}
