% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{pd_updateEstPars}
\alias{pd_updateEstPars}
\title{Update parameters}
\usage{
pd_updateEstPars(pd, parsEst, FLAGupdatePE = TRUE, FLAGsavePd = FALSE)
}
\arguments{
\item{FLAGupdatePE}{}
}
\value{

}
\description{
Update parameters
}
\seealso{
Other parframe handling: 
\code{\link{pd_parf_collectL1}()},
\code{\link{pd_parf_collectMstrust}()},
\code{\link{pd_parf_collectPars}()},
\code{\link{pd_parf_collectProfile}()},
\code{\link{pd_parf_collect}()},
\code{\link{readPd}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{parframe handling}
