% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_alignReplicates}
\alias{petab_alignReplicates}
\title{Use alignReplicates from bloIt on a PEtab}
\usage{
petab_alignReplicates(pe, i_subset, ...)
}
\arguments{
\item{pe}{petab}

\item{i_subset}{optional parameter for subsetting the dco with the i
parameter of data.table}

\item{...}{parameters for \link{blotIt::alignReplicates}}
}
\value{
\code{\link[=petab]{petab()}} with scaled replicates
}
\description{
The scaling acts on a PEtab by coercing it to a dco. The column names
passed \code{...} refer to the columns of the dco. The matching happens via
datapointId, so this column is mandatory.
}
\author{
Severin Bang (severin.bang@physik.uni-freiburg.de)
}
