% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_columns}
\alias{petab_columns}
\title{Title}
\usage{
petab_columns(pe = NULL)
}
\arguments{
\item{pe}{NULL: Default names, \code{\link[=petab]{petab()}} Actual names in petab}
}
\value{
list of column names
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
