% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_combine}
\alias{petab_combine}
\title{Title}
\usage{
petab_combine(
  pe1,
  pe2,
  NFLAGconflict = c(stop = 0, use_pe1 = 1, use_pe2 = 2)[2]
)
}
\arguments{
\item{NFLAGconflict}{one of c("stop" = 0, "use_pe1" = 1, "use_pe2" = 2)}
}
\value{

}
\description{
Title
}
\seealso{
Other Combine petabs: 
\code{\link{petab_combine_experimentalCondition}()},
\code{\link{petab_combine_measurementData}()},
\code{\link{petab_combine_observables}()},
\code{\link{petab_combine_parameters}()}
}
\concept{Combine petabs}
