% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_combine_experimentalCondition}
\alias{petab_combine_experimentalCondition}
\title{Title}
\usage{
petab_combine_experimentalCondition(ec1, ec2)
}
\arguments{
\item{ec2}{}
}
\value{

}
\description{
Title
}
