% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_combine_parameters}
\alias{petab_combine_parameters}
\title{Title}
\usage{
petab_combine_parameters(p1, p2)
}
\arguments{
\item{p2}{}
}
\value{

}
\description{
Title
}
\examples{

}
\seealso{
Other Combine petabs: 
\code{\link{petab_combine_experimentalCondition}()},
\code{\link{petab_combine_measurementData}()},
\code{\link{petab_combine_observables}()},
\code{\link{petab_combine}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Combine petabs}
