% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_create_parameter_df}
\alias{petab_create_parameter_df}
\title{Original petab function didn't work}
\usage{
petab_create_parameter_df(pe, observableParameterScale = "log10")
}
\arguments{
\item{measurementData}{}
}
\value{

}
\description{
Original petab function didn't work
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
