% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dModmodel2PE.R
\name{petab_dModmodel2PE}
\alias{petab_dModmodel2PE}
\title{Write PE from dMod objects}
\usage{
petab_dModmodel2PE(
  modelname,
  ODEmodel,
  obs_fun,
  errormodel,
  data,
  parameters,
  trafo,
  est_grid,
  fixed_grid,
  eventlist
)
}
\arguments{
\item{modelname}{}

\item{ODEmodel}{}

\item{obs_fun}{}

\item{errormodel}{}

\item{data}{DF with obligatory columns name, time, value, sigma, condition}

\item{parameters}{}

\item{trafo}{}

\item{eventlist}{}
}
\value{
pe list
}
\description{
Write PE from dMod objects
}
\author{
Svenja Kemmer
}
