% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{petab_examplePath}
\alias{petab_examplePath}
\title{Get paths of an example}
\usage{
petab_examplePath(exampleName, object = c("pe", "pd", "dir")[1])
}
\arguments{
\item{object}{}
}
\value{

}
\description{
Get paths of an example
}
\seealso{
Other Example functions: 
\code{\link{petab_exampleList}()},
\code{\link{petab_exampleRead}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Example functions}
