% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_getPars_estScale}
\alias{petab_getPars_estScale}
\title{Get parameters on outer scale}
\usage{
petab_getPars_estScale(pe)
}
\arguments{
\item{pe}{}
}
\value{

}
\description{
Get parameters on outer scale
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
