% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_joinDCO}
\alias{petab_joinDCO}
\title{Create one big table containing measurementData, observables and experimentalCondition}
\usage{
petab_joinDCO(pe, FLAGincludeMetaInformation = TRUE)
}
\arguments{
\item{petab}{\link{petab}}
}
\value{
data.table
}
\description{
conditionId is joined with simulationConditionId
}
\examples{
# Without metaInformation
pe <- petab_exampleRead("01")
dco <- petab_joinDCO(pe)
bla <- petab_unjoinDCO(dco,pe)

# With metaInformation including a "pattern" entry
pe <- petab_exampleRead("04")
pe <- petab_mutateDCO(pe, i = conditionId == "C1", j = `:=`(conditionId = "cell1_dose1"))
pe <- petab_mutateDCO(pe, i = conditionId == "C2", j = `:=`(conditionId = "cell1_dose2"))
pe$meta$metaInformation <- list(experimentalCondition = list(conditionId = list(pattern = "celltype_dose")))
dco <- petab_joinDCO(pe)
dco
}
\seealso{
Other dco: 
\code{\link{dco_expandPatterns}()},
\code{\link{petab_mutateDCO}()},
\code{\link{petab_unjoinDCO}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{dco}
