% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_joinDCO}
\alias{petab_joinDCO}
\title{Create one big table containing measurementData, observables and experimentalCondition}
\usage{
petab_joinDCO(pe, FLAGincludeMetaInformation = TRUE)
}
\arguments{
\item{petab}{\link{petab}}
}
\value{

}
\description{
Create one big table containing measurementData, observables and experimentalCondition
}
\examples{
pe <- petab_exampleRead("01")
dco <- petab_joinDCO(pe)
bla <- petab_unjoinDCO(dco,pe)
}
\seealso{
Other dco: 
\code{\link{dco_expandMetaColumns}()},
\code{\link{petab_mutateDCO}()},
\code{\link{petab_unjoinDCO}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{dco}
