% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_measurementData}
\alias{petab_measurementData}
\title{Constructor for Measurements}
\usage{
petab_measurementData(
  observableId,
  simulationConditionId,
  measurement,
  time,
  observableParameters = list(NA, "1;1", "scale_obsi;offset_obsi")[[1]],
  noiseParameters = list(1, "error_ADD_obsi;error_REL_obsi")[[1]],
  datasetId = NA,
  replicateId = NA,
  preequilibrationConditionId = NA
)
}
\arguments{
\item{observableParameters}{numeric string or NA}

\item{noiseParameters}{numeric, string or NA: Measurement noise or parameter name}
}
\description{
Constructor for Measurements
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
