% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_meta}
\alias{petab_meta}
\title{Additional Information}
\usage{
petab_meta(
  parameterFormulaInjection = NULL,
  variableOrder = NULL,
  metaInformation = NULL,
  ...
)
}
\arguments{
\item{parameterFormulaInjection}{data.table(parameterId, parameterFormula)}

\item{variableOrder}{vector of stateIds and observableIds in the order which you want for plotting}

\item{...}{}

\item{yaml}{hierarchichal list, stored in a yaml-file}
}
\value{
named list of supplied arguments
}
\description{
Additional Information
}
