% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_mutateDCO}
\alias{petab_mutateDCO}
\title{Consistently mutate the dco}
\usage{
petab_mutateDCO(pe, i, j)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}}}

\item{i}{expression to subset the dco, can be missing}

\item{j}{Must be a call of the form \verb{:=}(hsdhfgjs). If missing, pe is subsetted by i}
}
\value{
pe with modified DCO
}
\description{
If j is supplied, the full dco is always returned and no subsetting is done!
}
\examples{
# 1. supply i and j, mutate in rows defined by i. Important: Need to assign!
pe <- petab_exampleRead("02")
pe <- petab_mutateDCO(pe, replicateId == 2 & observableId == "obsE", `:=`(observableId = "obsE_special", observableFormula = "log(E)/log(7)"))
pe
# 2. supply i only. Subset to rows
pe <- petab_exampleRead("02")
pe <- petab_mutateDCO(pe, replicateId == 2 & observableId == "obsE")
pe
# 3. supply j only. Mutate all rows
pe <- petab_exampleRead("02")
pe <- petab_mutateDCO(pe, j = `:=`(observableId = "obsE_special", observableFormula = "log(E)/log(7)", observableName = "obsE_special", noiseFormula = "1"))
pe
}
\seealso{
Other dco: 
\code{\link{dco_expandPatterns}()},
\code{\link{petab_joinDCO}()},
\code{\link{petab_unjoinDCO}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{dco}
