% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_observables}
\alias{petab_observables}
\title{Constructor for Observables}
\usage{
petab_observables(
  observableId,
  observableName = NA,
  observableFormula = "observableParameter1_${observableId} * state1",
  observableTransformation = c("lin", "log", "log10")[[1]],
  noiseFormula = c(1,
    "noiseParameter${n}_${observableId} + noiseParameter${n}_${observableId}*${observableId}")[[1]],
  noiseDistribution = c("normal", "laplace")[[1]]
)
}
\arguments{
\item{observableId}{character}

\item{observableName}{character}

\item{observableFormula}{character}

\item{observableTransformation}{"lin", "log", "log10"}

\item{noiseFormula}{character}

\item{noiseDistribution}{"normal", "laplace"}
}
\value{
data.table
}
\description{
Constructor for Observables
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
