% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_overviewObsPerCond}
\alias{petab_overviewObsPerCond}
\title{Generate overview table which observables are in which condition}
\usage{
petab_overviewObsPerCond(
  pe,
  Ntruncate = 1000,
  FLAGincludedatasetId = TRUE,
  ...
)
}
\arguments{
\item{pe}{\code{\link[=petab]{petab()}} object}

\item{Ntruncate}{truncate pasted observables at this many characters}

\item{FLAGincludedatasetId}{summarize per conditionId, datasetId and replicateId}

\item{...}{arguments going to \code{\link[conveniencefunctions:cfoutput_MdTable]{conveniencefunctions::cfoutput_MdTable()}}}
}
\value{
prints table to console or writes it to disk
}
\description{
Generate overview table which observables are in which condition
}
\seealso{
Other Overview Tables: 
\code{\link{petab_overviewDCONames}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{Overview Tables}
