% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_parameters}
\alias{petab_parameters}
\title{Constructor for Parameters}
\usage{
petab_parameters(
  parameterId,
  parameterName = NA,
  parameterScale = c("log10", "log", "lin")[[1]],
  lowerBound = 1e-05,
  upperBound = 1000,
  nominalValue = 1,
  estimate = c(1, 0)[[1]],
  initializationPriorType = c("parameterScaleUniform", "uniform", "normal", "laplace",
    "logNormal", "logLaplace", "parameterScaleNormal", "parameterScaleLaplace")[[1]],
  initializationPriorParameters = "-1;1",
  objectivePriorType = c("parameterScaleNormal", "parameterScaleUniform", "uniform",
    "normal", "laplace", "logNormal", "logLaplace", "parameterScaleLaplace")[[1]],
  objectivePriorParameters = "0;2"
)
}
\arguments{
\item{objectivePriorParameters}{}
}
\value{

}
\description{
Constructor for Parameters
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
