% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_parameters_mergeParameters}
\alias{petab_parameters_mergeParameters}
\title{Title}
\usage{
petab_parameters_mergeParameters(
  pe_pa1,
  pe_pa2,
  mergeCols = c("nominalValue", "parameterScale", "estimate")
)
}
\arguments{
\item{pe_pa1}{parameters_df to merge values into}

\item{pe_pa2}{parameters_df to merge values from}

\item{mergeCols}{Columns you want to update in pe_pa1}
}
\value{
updated pe_pa1
}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
