% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_plotData}
\alias{petab_plotData}
\title{Plot Petab data}
\usage{
petab_plotData(
  petab,
  i,
  j,
  aeslist = NULL,
  FLAGUseObservableTransformation = TRUE,
  FLAGmeanLine = TRUE,
  ggCallback = list(facet_wrap_paginate(~observableId, nrow = 4, ncol = 4, scales =
    "free")),
  ...
)
}
\arguments{
\item{petab}{A \link{petab} object}

\item{aeslist}{list of formulas used to call \code{\link[ggplot2:aes_]{ggplot2::aes_q()}}.
Default aesthetics are list(x = ~time, y = ~measurement, color = ~conditionId).
Can be overriden by aeslist}

\item{FLAGmeanLine}{draw line connecting the means of groups defined by c("observableId", "time", "conditionId", names(aeslist)=}

\item{ggCallback}{additional stuff to add to the ggplot, such as a call to \code{\link[ggplot2:labs]{ggplot2::labs()}} or scales}

\item{FLAGfuture}{export asynchronously with the future package}

\item{filename, width, height, scale, units, ...}{see \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}}
}
\value{
ggplot
}
\description{
Automatic multipage export in the background with the future package and ggforce::facet_wrap_paginate
}
\seealso{
Other plotData: 
\code{\link{petab_duplicateControls}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotData}
