% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{petab_plotHelpers_meanMeasurementsValue}
\alias{petab_plotHelpers_meanMeasurementsValue}
\title{Title}
\usage{
petab_plotHelpers_meanMeasurementsValue(dplot, aeslist)
}
\arguments{
\item{aeslist}{}
}
\value{

}
\description{
Title
}
\seealso{
Other plotting: 
\code{\link{pd_plotParsParallelLines2}()},
\code{\link{pd_plotParsParallelLines}()},
\code{\link{pd_plotProfile}()},
\code{\link{petab_plotHelpers_aeslist}()},
\code{\link{petab_plotHelpers_parameterOrder}()},
\code{\link{petab_plotHelpers_variableOrder}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{plotting}
