% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{petab_python_setup}
\alias{petab_python_setup}
\title{Setup the connection to python petab}
\usage{
petab_python_setup(FLAGreturnpetabSelect = FALSE)
}
\arguments{
\item{FLAGreturnpetabSelect}{}
}
\value{
python module, see \code{\link[reticulate:import]{reticulate::import()}}
}
\description{
\itemize{
\item if petab virtualenv not present:
\itemize{
\item sets up a virtualenv called "petab"
\item pip installs petab
}
\item uses "petab" virtual env
\item imports and returns petab
}
}
\details{
use as pe <- petab_python_setup()
}
\examples{
pepy <- petab_python_setup()
# pepy$lint(pe)

peps <- petab_python_setup(FLAGreturnpetabSelect = TRUE)
}
\seealso{
Other pepy: 
\code{\link{pepy_sample_parameter_startpoints}()},
\code{\link{petab_python_reinstall}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{pepy}
