% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petab_select.R
\name{petabSelect_collectReportYamlCriteria}
\alias{petabSelect_collectReportYamlCriteria}
\title{Ensure consistent naming of Criteria}
\usage{
petabSelect_collectReportYamlCriteria(AIC, BIC, AICc)
}
\arguments{
\item{AIC, BIC, AICc}{numeric(1L), the respective criteria}
}
\value{
list(AIC, BIC, AICc)
}
\description{
Ensure consistent naming of Criteria
}
\seealso{
Other petab select: 
\code{\link{as.data.table.reportYaml}()},
\code{\link{pd_petabSelect_collectEstimatedParameters}()},
\code{\link{pd_petabSelect_collectParameters}()},
\code{\link{pd_petabSelect_collectReportYamlCriteria}()},
\code{\link{pd_petabSelect_reportYaml}()},
\code{\link{petabSelect_LRT}()},
\code{\link{petabSelect_compareModels}()},
\code{\link{petabSelect_readReportYaml}()},
\code{\link{petabSelect_reportYaml}()},
\code{\link{petabSelect_writeReportYaml}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{petab select}
