% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd.R
\name{readPd}
\alias{readPd}
\title{Read a pd and load dlls}
\usage{
readPd(filename)
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Read a pd and load dlls
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
