% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sanitizePowers}
\alias{sanitizePowers}
\title{Sanitize pow(x,y) equations}
\usage{
sanitizePowers(rate)
}
\arguments{
\item{rate}{}
}
\value{

}
\description{
Sanitize pow(x,y) equations
}
\examples{
sanitizePowers("EMAX * pow(Ac / Vc + eps, hill) / (pow(Ac / Vc + eps, hill) + pow(EC50, hill)) * cytoplasm")
rate <- "EMAX * pow(Ac /   Vc, hill+2)"
sanitizePowers(rate)
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{sbmlImport}
