% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBML-Export.R
\name{sbml_exportEquationList}
\alias{sbml_exportEquationList}
\title{Write sbml model to filename}
\usage{
sbml_exportEquationList(
  equationList,
  filename,
  modelname = "Model",
  unitInfo = getUnitInfo(),
  speciesInfo = getSpeciesInfo(equationList),
  parInfo = getParInfo(equationList, eventList = events),
  compartmentInfo = getCompartmentInfo(equationList),
  events = NULL
)
}
\arguments{
\item{equationList}{}

\item{filename}{}

\item{modelname}{}

\item{unitInfo}{}

\item{speciesInfo}{}

\item{parInfo}{}

\item{compartmentInfo}{}
}
\value{
NULL, called for side effect
}
\description{
Write sbml model to filename
}
\examples{
equationList <- NULL
equationList <- addReaction(equationList, from = "E + S", to = "ES", rate = "(kon)*E*S",
                            description = "production of complex")
equationList <- addReaction(equationList, from = "ES", to = "E + S", rate = "koff*ES",
                            description = "decay of complex")
equationList <- addReaction(equationList, from = "ES", to = "E + P", rate = "kcat*ES",
                            description = "production of product")
equationList <- eqnlist_addDefaultCompartment(equationList, "cytoplasm")

filename <- modelname <- file.path(tempdir(), "model.xml")

parInfo <- data.table(tibble::tribble(
  ~parName, ~parValue, ~parUnit,
  "kon"   ,     0.10,"litre_per_mole_per_second" ,
  "koff"  ,     0.55,"per_second" ,
  "kcat"  ,     1.00,"per_second" ))
unitInfo <- getUnitInfo()
compartmentInfo <- getCompartmentInfo(equationList)
speciesInfo <- getSpeciesInfo(equationList)

# Eventlist 1 with values only
eventList <- eventlist(var = "E", time = 0, value = 1, root = NA, method = "replace")
eventList <- addEvent(eventList, var = "E", time = 1, value = 0, root = NA, method = "replace")
eventInfo <- getEventInfo(eventList)

# Eventlist 2 with parameters as well
eventList2 <- eventlist(var = "E", time = 0, value = "E_event1", root = NA, method = "replace")
eventList2 <- addEvent(eventList2, var = "E", time = 1, value = "E_event2", root = NA, method = "replace")
eventInfo2 <- getEventInfo(eventList2)
parInfo2 <- getParInfo(equationList, eventList2)


sbml_exportEquationList(equationList, filename, parInfo = parInfo)
sbml_exportEquationList(equationList, filename, parInfo = parInfo, eventList = eventList)
sbml_exportEquationList(equationList, filename, parInfo = parInfo2, eventList = eventList2) # don't know why this throws an error..
}
\seealso{
Other SBML export: 
\code{\link{eqnlist_addDefaultCompartment}()},
\code{\link{getCompartmentInfo}()},
\code{\link{getParInfo}()},
\code{\link{getReactionInfo}()},
\code{\link{getSpeciesInfo}()},
\code{\link{getUnitInfo}()},
\code{\link{petab_dModmodel2PE}()},
\code{\link{sbml_addOneCompartment}()},
\code{\link{sbml_addOneEvent}()},
\code{\link{sbml_addOneParameter}()},
\code{\link{sbml_addOneReaction}()},
\code{\link{sbml_addOneSpecies}()},
\code{\link{sbml_addOneUnit}()},
\code{\link{sbml_initialize}()},
\code{\link{sbml_kineticLawAddParameters}()},
\code{\link{sbml_reactionAddKineticLaw}()},
\code{\link{sbml_reactionAddModifiers}()},
\code{\link{sbml_reactionAddProducts}()},
\code{\link{sbml_reactionAddReactants}()},
\code{\link{sbml_validateSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML export}
