% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sbmlImport_getParametersFromSBML}
\alias{sbmlImport_getParametersFromSBML}
\title{Extract parameters from SBML}
\usage{
sbmlImport_getParametersFromSBML(modelfile)
}
\arguments{
\item{modelfile}{Path to SBML file}
}
\value{
\code{\link[=petab_parameters]{petab_parameters()}} with scale = "lin" and estimate = 0
}
\description{
Assumes the parameters are given numerically
}
\examples{
modelfile <- file.path(petab_examplePath("01", "pe"), "model_petab.xml")
sbmlImport_getCompartmentsFromSBML(modelfile)
}
\seealso{
Other SBML import: 
\code{\link{sbmlImport_getCompartmentsFromSBML}()}
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
\concept{SBML import}
