% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-import.R
\name{sbmlImport_getReactionDetails}
\alias{sbmlImport_getReactionDetails}
\title{Title}
\usage{
sbmlImport_getReactionDetails(m, reaction, compartments)
}
\arguments{
\item{m}{SBML model}

\item{reaction}{index of reaction}

\item{compartments}{names of compartments}
}
\value{

}
\description{
import reactions and adjust by means of compartments
}
