% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{subsetPredictionToData}
\alias{subsetPredictionToData}
\title{Subset predictions to observables and conditions observed in data}
\usage{
subsetPredictionToData(
  pplot,
  dplot,
  NFLAGsubsetType = c(none = 0, strict = 1, keepInternal = 2, strict_cutTimes = 3,
    keepInternal_cutTimes = 3)[2]
)
}
\arguments{
\item{NFLAGsubsetType}{}
}
\value{

}
\description{
Subset predictions to observables and conditions observed in data
}
