% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMod-pd.R
\name{subsetPredictionToData}
\alias{subsetPredictionToData}
\title{Subset predictions to observables and conditions observed in data}
\usage{
subsetPredictionToData(pplot, dplot)
}
\arguments{
\item{dplot}{}
}
\value{

}
\description{
Subset predictions to observables and conditions observed in data
}
