% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEtab.R
\name{writePetab}
\alias{writePetab}
\title{Title}
\usage{
writePetab(pe, filename = "petab/model")
}
\arguments{
\item{filename}{}
}
\value{

}
\description{
Title
}
\author{
Daniel Lill (daniel.lill@physik.uni-freiburg.de)
}
